### 2020-01-27 220398a
Make cost 10 for mv_mtaw in docs
