+++ venv/pyvenv.cfg
@@ -1,0 +1,3 @@
+home = D:\Program Files\Anaconda
+include-system-site-packages = false
+version = 3.7.3
