### 2020-01-27 9fd15d6
cleaning up left-over tests
