+++ tests/test_search_itp_quartairestratigrafie.py
@@ -1,197 +1,0 @@
-"""Module grouping tests for the interpretaties search module."""
-import pandas as pd
-import numpy as np
-import pytest
-from pandas import DataFrame
-
-import pydov
-from owslib.fes import PropertyIsEqualTo
-from pydov.search.interpretaties import QuartaireStratigrafieSearch
-from pydov.types.interpretaties import QuartaireStratigrafie
-from tests.abstract import (
-    AbstractTestSearch,
-)
-
-from tests.test_search import (
-    mp_wfs,
-    wfs,
-    mp_remote_md,
-    mp_remote_fc,
-    mp_remote_describefeaturetype,
-    mp_remote_wfs_feature,
-    mp_dov_xml,
-    mp_dov_xml_broken,
-    wfs_getfeature,
-    wfs_feature,
-)
-
-location_md_metadata = \
-    'tests/data/types/interpretaties/quartaire_stratigrafie/md_metadata.xml'
-location_fc_featurecatalogue = \
-    'tests/data/types/interpretaties/quartaire_stratigrafie/fc_featurecatalogue.xml'
-location_wfs_describefeaturetype = \
-    'tests/data/types/interpretaties/quartaire_stratigrafie/wfsdescribefeaturetype.xml'
-location_wfs_getfeature = \
-    'tests/data/types/interpretaties/quartaire_stratigrafie/wfsgetfeature.xml'
-location_wfs_feature = \
-    'tests/data/types/interpretaties/quartaire_stratigrafie/feature.xml'
-location_dov_xml = \
-    'tests/data/types/interpretaties/quartaire_stratigrafie/quartaire_stratigrafie.xml'
-
-
-class TestQuartaireStratigrafieSearch(AbstractTestSearch):
-    def get_search_object(self):
-        """Get an instance of the search object for this type.
-
-        Returns
-        -------
-        pydov.search.interpretaties.QuartaireStratigrafieSearch
-            Instance of QuartaireStratigrafieSearch used for searching.
-
-        """
-        return QuartaireStratigrafieSearch()
-
-    def get_type(self):
-        """Get the class reference for this datatype.
-
-        Returns
-        -------
-        pydov.types.interpretaties.QuartaireStratigrafie
-            Class reference for the QuartaireStratigrafie class.
-
-        """
-        return QuartaireStratigrafie
-
-    def get_valid_query_single(self):
-        """Get a valid query returning a single feature.
-
-        Returns
-        -------
-        owslib.fes.OgcExpression
-            OGC expression of the query.
-
-        """
-        return PropertyIsEqualTo(propertyname='Proefnummer',
-                                 literal='GEO-42/190-B14')
-
-    def get_inexistent_field(self):
-        """Get the name of a field that doesn't exist.
-
-        Returns
-        -------
-        str
-            The name of an inexistent field.
-
-        """
-        return 'onbestaand'
-
-    def get_xml_field(self):
-        """Get the name of a field defined in XML only.
-
-        Returns
-        -------
-        str
-            The name of the XML field.
-
-        """
-        return 'diepte_laag_van'
-
-    def get_valid_returnfields(self):
-        """Get a list of valid return fields from the main type.
-
-        Returns
-        -------
-        tuple
-            A tuple containing only valid return fields.
-
-        """
-        return ('pkey_interpretatie', 'betrouwbaarheid_interpretatie')
-
-    def get_valid_returnfields_subtype(self):
-        """Get a list of valid return fields, including fields from a subtype.
-
-        Returns
-        -------
-        tuple
-            A tuple containing valid return fields, including fields from a
-            subtype.
-
-        """
-        return ('pkey_interpretatie', 'diepte_laag_van', 'diepte_laag_tot')
-
[…]
