### 2020-01-27 90e94e1
Fix #86 by adding documentation on dev environment
