### 2020-01-27 8c038f2
Remove coveralls from repo
