### 2020-01-27 2d2b022
Revert "Make cost 10 for mv_mtaw in docs"
This reverts commit 220398a918bfe5878e7f233b90753a629c8e965d.
