+++ venv/pyvenv.cfg
@@ -1,3 +1,0 @@
-home = D:\Program Files\Anaconda
-include-system-site-packages = false
-version = 3.7.3
