### 2020-01-27 cf83bd5
Make cost 10 for mv_mtaw in docs.
