### 2020-01-27 a6bbbd2
added explanation in output_fields.rst and added example in notebook
