### 2020-01-27 905439e
added explanation in output_fields.rst and added example in notebook
