### 2020-01-27 2ce7530
also remove pins in appveyor
