### 2020-01-27 593149f
Update get_remote_metadata and get_featurecatalogue_uuid.
