+++ pydov/types/grondwaterfilter.py
@@ -87,7 +87,8 @@ class GrondwaterFilter(AbstractDovType):
         XmlField(name='mv_mtaw',
                  source_xpath='/grondwaterlocatie/puntligging/'
                               'oorspronkelijk_maaiveld/waarde',
-                 definition='Maaiveldhoogte in mTAW op dag dat de put/boring uitgevoerd werd',
+                 definition='Maaiveldhoogte in mTAW op dag '
+                            'dat de put/boring uitgevoerd werd',
                  datatype='float'),
         WfsField(name='gemeente', source_field='gemeente', datatype='string'),
         XmlField(name='meetnet_code',
