### 2020-01-28 2634675
Remove coveralls from repo (#230)
