+++ README.md
@@ -1,6 +1,6 @@
 # pydov <img src="docs/_static/img/logo.png" align="right" alt="" width="120">
 
-[![Build Status](https://travis-ci.org/DOV-Vlaanderen/pydov.svg?branch=master)](https://travis-ci.org/DOV-Vlaanderen/pydov) [![AppVeyor Build status](https://ci.appveyor.com/api/projects/status/4ljy2a0p661v3d9k/branch/master?svg=true)](https://ci.appveyor.com/project/Roel/pydov) [![Documentation Status](https://readthedocs.org/projects/pydov/badge/?version=latest)](https://pydov.readthedocs.io/en/latest/?badge=latest) [![Coverage Status](https://coveralls.io/repos/github/DOV-Vlaanderen/pydov/badge.svg?branch=master)](https://coveralls.io/github/DOV-Vlaanderen/pydov?branch=master) [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2788680.svg)](https://doi.org/10.5281/zenodo.2788680)
+[![Build Status](https://travis-ci.org/DOV-Vlaanderen/pydov.svg?branch=master)](https://travis-ci.org/DOV-Vlaanderen/pydov) [![AppVeyor Build status](https://ci.appveyor.com/api/projects/status/4ljy2a0p661v3d9k/branch/master?svg=true)](https://ci.appveyor.com/project/Roel/pydov) [![Documentation Status](https://readthedocs.org/projects/pydov/badge/?version=latest)](https://pydov.readthedocs.io/en/latest/?badge=latest) [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2788680.svg)](https://doi.org/10.5281/zenodo.2788680)
 
 pydov is a Python package to query and download data from [Databank Ondergrond Vlaanderen (DOV)](https://www.dov.vlaanderen.be). It is hosted on [GitHub](https://github.com/DOV-Vlaanderen/pydov) and development is coordinated by Databank Ondergrond Vlaanderen (DOV). DOV aggregates data about soil, subsoil and groundwater of Flanders and makes them publicly available. Interactive and human-readable extraction and querying of the data is provided by a [web application](https://www.dov.vlaanderen.be/portaal/?module=verkenner#ModulePage), whereas the focus of this package is to **support machine-based extraction and conversion of the data**.
 
