### 2020-01-28 5dd09d9
Added mv_mtaw to grondwaterfilter type (#227)
* Added mv_mtaw to grondwaterfilter type

* Make cost 10 for mv_mtaw in docs

* Modify notebook.
