### 2020-01-28 9d31a86
Change documentation to be in line with code example
