### 2020-01-28 e8055d1
Add Python 3.8 as supported version
