### 2020-01-28 f0d807b
Remove Python 2 workaround for unicode
