### 2020-01-28 c545317
Add Joris Synaeve to Author list.
