+++ AUTHORS.rst
@@ -8,3 +8,4 @@ Credits
 * Stijn Van Hoey
 * Johan Van de Wauw
 * Roel Huybrechts <roel.huybrechts@realdolmen.com>
+* Joris Synaeve
