### 2020-01-28 d589beb
Fix typo's/improve readability
Remove some non-existing words.
