### 2020-01-28 bf2ef8e
Remove obsolete output data description
