### 2020-01-28 a4a6e42
Update bounding box (#232)
* changed quickstart.rst with a warning of the large data request resulting from the second example with a large bounding box

* changed quickstart to use smaller bounding box

* Change documentation to be in line with code example

* Edit boudingbox in readme.md.

* Fix readme boundingbox.

Co-authored-by: jorissynaeve <38761924+jorissynaeve@users.noreply.github.com>
