+++ docs/quickstart.rst
@@ -34,7 +34,7 @@ Each of the datasets can be queried using a search object for this dataset. Whil
     from pydov.search.boring import BoringSearch
     boringsearch = BoringSearch()
 
-Now we can query for boreholes either on attributes, on location or on a combination of both. To query on attributes, we use the OGC filter functions from OWSLib. For example, to request all boreholes with a depth over 2000 m, we would use the following ``query`` parameter:
+Now we can query for boreholes either on attributes, on location or on a combination of both. To query on attributes, we use the OGC filter functions from OWSLib. For example, to request all boreholes with a depth over 550 m, we would use the following ``query`` parameter:
 
 ::
 
@@ -42,7 +42,7 @@ Now we can query for boreholes either on attributes, on location or on a combina
 
     dataframe = boringsearch.search(
         query=PropertyIsGreaterThan(
-            propertyname='diepte_tot_m', literal='2000')
+            propertyname='diepte_tot_m', literal='550')
     )
 
 To query on location, we use location objects and spatial filters from the pydov.util.location module. For example, to request all boreholes in a given bounding box, we would use the following ``location`` parameter:
@@ -52,7 +52,7 @@ To query on location, we use location objects and spatial filters from the pydov
     from pydov.util.location import Within, Box
 
     dataframe = boringsearch.search(
-        location=Within(Box(94720, 186910, 112220, 202870))
+        location=Within(Box(107500, 202000, 108500, 203000))
     )
 
 Attribute queries can be combined with location filtering by specifying both parameters in the search call:
@@ -61,8 +61,8 @@ Attribute queries can be combined with location filtering by specifying both par
 
     dataframe = boringsearch.search(
         query=PropertyIsGreaterThan(
-            propertyname='diepte_tot_m', literal='2000'),
-        location=Within(Box(94720, 186910, 112220, 202870))
+            propertyname='diepte_tot_m', literal='550'),
+        location=Within(Box(107500, 202000, 108500, 203000))
     )
 
 The :ref:`query_attribute` and :ref:`query_location` pages provide an overview of the query options for attributes and locations respectively.
