### 2020-01-28 6401e4e
Fix output dataframe fields for groundwater samples.
