### 2020-01-28 b87b6d8
Move explanation to GrondwaterFilter datatype.
It is the only where this is applicable.
