+++ AUTHORS.rst
@@ -7,4 +7,4 @@ Credits
 * Pieter Jan Haest <pieterjan.haest@agt.be>
 * Stijn Van Hoey
 * Johan Van de Wauw
-* Roel Huybrechts <roel.huybrechts@realdolmen.com>
+* Roel Huybrechts <roel@huybrechts.re>
