### 2020-01-29 3ba60be
Add description for datum and tijd (rebased) (#235)
* added explanation in output_fields.rst and added example in notebook

* Fix output dataframe fields for groundwater samples.

* Move explanation to GrondwaterFilter datatype.

It is the only where this is applicable.

* Readd example of combining date and time.

Co-authored-by: Roel Huybrechts <roel@huybrechts.re>
