### 2020-01-29 a5e686c
Remove some code duplication between pydov and owslib (#238)
* Remove copied get_schema method from OWSLib.

* Update get_remote_metadata and get_featurecatalogue_uuid.

* Fix offline tests.

* Remove unused imports.
