+++ tests/abstract.py
@@ -217,8 +217,9 @@ class AbstractTestSearch(object):
         datatype = self.get_type()
         self.get_search_object().__class__(objecttype=datatype)
 
-    def test_get_fields(self, mp_wfs, mp_remote_describefeaturetype,
-                        mp_remote_md, mp_remote_fc, mp_remote_xsd):
+    def test_get_fields(self, mp_wfs, mp_get_schema,
+                        mp_remote_describefeaturetype, mp_remote_md,
+                        mp_remote_fc, mp_remote_xsd):
         """Test the get_fields method.
 
         Test whether the returned fields match the format specified
@@ -228,6 +229,8 @@ class AbstractTestSearch(object):
         ----------
         mp_wfs : pytest.fixture
             Monkeypatch the call to the remote GetCapabilities request.
+        mp_get_schema : pytest.fixture
+            Monkeypatch the call to a remote OWSLib schema.
         mp_remote_describefeaturetype : pytest.fixture
             Monkeypatch the call to a remote DescribeFeatureType.
         mp_remote_md : pytest.fixture
@@ -293,7 +296,8 @@ class AbstractTestSearch(object):
                 assert sorted(f.keys()) == ['cost', 'definition', 'name',
                                             'notnull', 'query', 'type']
 
-    def test_search_both_location_query(self, mp_remote_describefeaturetype,
+    def test_search_both_location_query(self, mp_get_schema,
+                                        mp_remote_describefeaturetype,
                                         mp_remote_wfs_feature):
         """Test the search method providing both a location and a query.
 
@@ -301,6 +305,8 @@ class AbstractTestSearch(object):
 
         Parameters
         ----------
+        mp_get_schema : pytest.fixture
+            Monkeypatch the call to a remote OWSLib schema.
         mp_remote_describefeaturetype : pytest.fixture
             Monkeypatch the call to a remote DescribeFeatureType.
         mp_remote_wfs_feature : pytest.fixture
@@ -314,8 +320,9 @@ class AbstractTestSearch(object):
 
         assert type(df) is DataFrame
 
-    def test_search(self, mp_wfs, mp_remote_describefeaturetype, mp_remote_md,
-                    mp_remote_fc, mp_remote_wfs_feature, mp_dov_xml):
+    def test_search(self, mp_wfs, mp_get_schema, mp_remote_describefeaturetype,
+                    mp_remote_md, mp_remote_fc, mp_remote_wfs_feature,
+                    mp_dov_xml):
         """Test the search method with only the query parameter.
 
         Test whether the result is correct.
@@ -324,6 +331,8 @@ class AbstractTestSearch(object):
         ----------
         mp_wfs : pytest.fixture
             Monkeypatch the call to the remote GetCapabilities request.
+        mp_get_schema : pytest.fixture
+            Monkeypatch the call to a remote OWSLib schema.
         mp_remote_describefeaturetype : pytest.fixture
             Monkeypatch the call to a remote DescribeFeatureType.
         mp_remote_md : pytest.fixture
@@ -503,13 +512,16 @@ class AbstractTestSearch(object):
         with pytest.raises(InvalidFieldError):
             self.get_search_object().search(query=query)
 
-    def test_search_extrareturnfields(self, mp_remote_describefeaturetype,
+    def test_search_extrareturnfields(self, mp_get_schema,
+                                      mp_remote_describefeaturetype,
                                       mp_remote_wfs_feature, mp_dov_xml):
         """Test the search method with the query parameter and an extra WFS
         field as return field.
 
         Parameters
         ----------
+        mp_get_schema : pytest.fixture
+            Monkeypatch the call to a remote OWSLib schema.
[…]
