### 2020-01-30 4abfbe4
Tests: use HEAD instead of GET in service_ok
Since we only rely on the status code and the headers to determine if a
service is working or not, it is not needed to perform a GET (resulting
in downloads of data we don't need).

Instead we could use a HEAD request. Since Requests does not follow
redirects by default when using HEAD [1], we enable it explicitly using
allow_redirects.

Also update to use the global pydov requests session instead of creating
a new one for each call.

[1]: https://2.python-requests.org/en/master/user/quickstart/#redirection-and-history
