+++ pydov/util/dovutil.py
@@ -59,7 +59,18 @@ def get_xsd_schema(url):
         The raw XML data of this XSD schema as bytes.
 
     """
-    return get_remote_url(url)
+    response = None
+    for hook in pydov.hooks:
+        r = hook.inject_meta_response(url)
+        if r is not None:
+            response = r.encode('utf8')
+
+    if response is None:
+        response = get_remote_url(url)
+
+    for hook in pydov.hooks:
+        hook.meta_received(url, response)
+    return response
 
 
 def get_dov_xml(url):
@@ -76,7 +87,19 @@ def get_dov_xml(url):
         The raw XML data of this DOV object as bytes.
 
     """
-    return get_remote_url(url)
+    response = None
+    for hook in pydov.hooks:
+        r = hook.inject_xml_retrieved(url)
+        if r is not None:
+            response = r.encode('utf8')
+
+    if response is None:
+        response = get_remote_url(url)
+
+    for hook in pydov.hooks:
+        hook.xml_retrieved(url, response)
+
+    return response
 
 
 def parse_dov_xml(xml_data):
