+++ pydov/search/abstract.py
@@ -117,8 +117,28 @@ class AbstractSearch(AbstractCommon):
         to all subclasses and instances.
         """
         if AbstractSearch.__wfs is None:
-            AbstractSearch.__wfs = WebFeatureService(
-                url=build_dov_url('geoserver/wfs'), version="1.1.0")
+            capabilities = None
+
+            for hook in pydov.hooks:
+                capa = hook.inject_meta_response(
+                    build_dov_url('geoserver/wfs') + '?version=1.1.0')
+                if capa is not None:
+                    capabilities = capa
+
+            if capabilities is None:
+                AbstractSearch.__wfs = WebFeatureService(
+                    url=build_dov_url('geoserver/wfs'), version="1.1.0")
+            else:
+                AbstractSearch.__wfs = WebFeatureService(
+                    url=build_dov_url('geoserver/wfs'), version="1.1.0",
+                    xml=capabilities)
+
+            for hook in pydov.hooks:
+                hook.meta_received(
+                    build_dov_url('geoserver/wfs') + '?version=1.1.0',
+                    etree.tostring(AbstractSearch.__wfs._capabilities,
+                                   encoding='utf8')
+                )
 
     def _init_namespace(self):
         """Initialise the WFS namespace associated with the layer.
@@ -529,7 +549,7 @@ class AbstractSearch(AbstractCommon):
 
         Returns
         -------
-        bytes
+        wfs_response, wfs_getfeature_request : bytes, etree.Element
             Response of the WFS service.
 
         """
@@ -546,11 +566,21 @@ class AbstractSearch(AbstractCommon):
 
         for hook in pydov.hooks:
             hook.wfs_search_init(typename)
+            hook.wfs_search_query(wfs_getfeature_xml)
+
+        tree = None
+        for hook in pydov.hooks:
+            t = hook.inject_wfs_result_features(wfs_getfeature_xml)
+            if t is not None:
+                tree = t
+
+        if tree is not None:
+            return tree, wfs_getfeature_xml
 
         return owsutil.wfs_get_feature(
             baseurl=wfs.url,
             get_feature_request=wfs_getfeature_xml
-        )
+        ), wfs_getfeature_xml
 
     def _search(self, location=None, query=None, return_fields=None,
                 sort_by=None, max_features=None, extra_wfs_fields=[]):
@@ -646,7 +676,7 @@ class AbstractSearch(AbstractCommon):
 
             sort_by = etree.tostring(sort_by_xml, encoding='unicode')
 
-        fts = self._get_remote_wfs_feature(
+        fts, getfeature = self._get_remote_wfs_feature(
             wfs=self.__wfs,
             typename=self._layer,
             location=location,
@@ -670,6 +700,7 @@ class AbstractSearch(AbstractCommon):
 
         for hook in pydov.hooks:
             hook.wfs_search_result(int(tree.get('numberOfFeatures')))
+            hook.wfs_search_result_features(getfeature, tree)
 
         return tree
 
