+++ pydov/util/owsutil.py
@@ -461,7 +461,18 @@ def get_url(url):
         Response containing the result of the GET request.
 
     """
+    response = None
+    for hook in pydov.hooks:
+        r = hook.inject_meta_response(url)
+        if r is not None:
+            response = r.encode('utf8')
 
-    request = pydov.session.get(url, timeout=pydov.request_timeout)
-    request.encoding = 'utf-8'
-    return request.text.encode('utf8')
+    if response is None:
+        request = pydov.session.get(url, timeout=pydov.request_timeout)
+        request.encoding = 'utf-8'
+        response = request.text.encode('utf8')
+
+    for hook in pydov.hooks:
+        hook.meta_received(url, response)
+
+    return response
