### 2020-01-30 b948414
Generate stable WFS GetFeature requests.
The WFS GetFeature builder should make the same request when given the
same input.
