### 2020-01-30 1bd5cf4
Add test for stable GetFeature propertynames.
Fix existing tests too.
