### 2020-01-31 f5d985b
Add docstrings and remove two hooks.
Remove wfs_search_query and xml_requested hooks.
