### 2020-01-31 c49925d
More consistent hook naming.
