+++ tests/test_util_hooks.py
@@ -1,8 +1,11 @@
+import copy
 import datetime
 import pytest
 
 import pydov
 from owslib.fes import PropertyIsEqualTo
+from owslib.etree import etree
+from pydov.search.abstract import AbstractSearch
 from pydov.search.boring import BoringSearch
 from pydov.util.hooks import (
     AbstractHook,
@@ -14,6 +17,52 @@ from tests.test_util_caching import (
     plaintext_cache,
     nocache,
 )
+
+
+@pytest.fixture
+def test_hook_count():
+    """PyTest fixture temporarily disabling default hooks and installing
+    HookCounter."""
+    orig_hooks = pydov.hooks
+    temp_hooks = [HookCounter()]
+
+    pydov.hooks = temp_hooks
+    yield
+
+    pydov.hooks = orig_hooks
+
+
+@pytest.fixture
+def test_hook_types():
+    """PyTest fixture temporarily disabling default hooks and installing
+    HookTester."""
+    orig_hooks = pydov.hooks
+    temp_hooks = [HookTypeTester()]
+
+    pydov.hooks = temp_hooks
+    yield
+
+    pydov.hooks = orig_hooks
+
+
+@pytest.fixture
+def test_hook_inject():
+    """PyTest fixture temporarily disabling default hooks and installing
+    HookInjecter."""
+    orig_hooks = pydov.hooks
+    temp_hooks = [HookInjecter()]
+
+    pydov.hooks = temp_hooks
+    yield
+
+    pydov.hooks = orig_hooks
+
+
+@pytest.fixture
+def force_rebuild_wfs():
+    """PyTest fixture to temporarily reset the cached WFS instance to force
+    reevaluation of the initialisation code and recall meta events."""
+    AbstractSearch._AbstractSearch__wfs = None
 
 
 class HookCounter(AbstractHook):
@@ -61,35 +110,118 @@ class HookCounter(AbstractHook):
         self.count_xml_downloaded += 1
 
 
-@pytest.fixture
-def temp_hooks():
-    """PyTest fixture temporarily disabling default hooks and installing
-    HookCounter."""
-    orig_hooks = pydov.hooks
-    temp_hooks = [HookCounter()]
+class HookTypeTester(AbstractHook):
+    """Hook implementation for testing purposes, testing arguments of all
+    event calls."""
+    def meta_received(self, url, response):
+        assert url is not None
+        assert response is not None
+        assert type(url) is str
+        assert type(response) is bytes
 
-    pydov.hooks = temp_hooks
-    yield
+    def inject_meta_response(self, url):
+        assert url is not None
+        assert type(url) is str
 
-    pydov.hooks = orig_hooks
+    def wfs_search_init(self, typename):
+        assert typename is not None
+        assert type(typename) is str
+
+    def wfs_search_result(self, number_of_results):
+        assert number_of_results is not None
+        assert number_of_results > 0
+        assert type(number_of_results) is int
+
+    def wfs_search_result_received(self, query, features):
+        assert query is not None
+        assert features is not None
+
+    def inject_wfs_getfeature_response(self, query):
+        assert query is not None
+
+    def xml_received(self, pkey_object, xml):
+        assert pkey_object is not None
+        assert xml is not None
+        assert type(pkey_object) is str
+        assert type(xml) is bytes
+
+    def inject_xml_response(self, pkey_object):
+        assert pkey_object is not None
+        assert type(pkey_object) is str
+
+    def xml_cache_hit(self, pkey_object):
+        assert pkey_object is not None
+        assert type(pkey_object) is str
+
+    def xml_downloaded(self, pkey_object):
+        assert pkey_object is not None
+        assert type(pkey_object) is str
 
 
-class TestHooks(object):
[…]
