+++ docs/hooks.rst
@@ -44,24 +44,71 @@ implementation of the hooks itself.
 Available event hooks
 .....................
 
-wfs_search_init
+meta_received (url: str, response: bytes)
+    This method will be called whenever a response for a metadata request is
+    received. There are two parameters, `url` with the full URL of the metadata
+    request and `response` with the response from the server.
+
+    Metadata requests include, amongst others: WFS GetCapabilities,
+    requests for MD_Metadata, FC_FeatureCatalogue and XSD schemas. These are
+    all calls except for WFS GetFeature requests and XML downloads of DOV data
+    - these are other hooks.
+
+inject_meta_response (url: str) -> bytes
+    This method can be used to inject a custom response for a metadata request
+    with the given URL. There is one parameter `url` with the full URL of the
+    metadata request.
+
+    When at least one registered hook returns a response for a given URL,
+    the remote call is not executed and instead the response from the
+    last registered hook (that is non-null) is used instead.
+
+wfs_search_init (typename: str)
     This method will be called whenever a WFS search is initiated. There is
     one parameter `typename` with the WFS typename that is queried.
 
-wfs_search_result
+wfs_search_result (number_of_results: int)
     This method will be called whenever a WFS search is completed. There is
     one parameter `number_of_results` with the number of search results.
 
-xml_requested
-    This method will be called whenever an XML document is needed. There is
-    one parameter `pkey_object` with the permanent key of the DOV object.
-    This event is either followed by `xml_cache_hit` or `xml_downloaded`.
+wfs_search_result_received (query: etree.ElementTree, features: etree.ElementTree)
+    This method will be called whenever a WFS search finished. There are two
+    parameters, `query` is the WFS GetFeature request sent to the server and
+    `features` is the FeatureCollection received in response.
+
+inject_wfs_getfeature_response (query: etree.ElementTree) -> bytes
+    This method can be used to inject a custom response for a WFS GetFeature
+    request with the given query. There is one parameter `query` with the WFS
+    GetFeature request sent to the server.
+
+    When at least one registered hook returns a response for a given query,
+    the remote call is not executed and instead the response from the
+    last registered hook (that is non-null) is used instead.
+
+xml_received (pkey_object: str, xml: bytes)
+    This method will be called whenever an XML document is received, either
+    from the cache or from the remote DOV service. There are two parameters,
+    `pkey_object` with the permanent key of the DOV object and `xml` containing
+    the full XML representation.
 
     Because of parallel processing, this method will be called simultaneously
     from multiple threads. Make sure your implementation is threadsafe or uses
     locking.
 
-xml_cache_hit
+inject_xml_response (pkey_object: str) -> bytes
+    This method can be used to inject a custom response for a DOV XML
+    request for the given object. There is one parameter `pkey_object` with
+    the permanent key of the DOV object.
+
+    When at least one registered hook returns a response for a given pkey,
+    the remote call is not executed and instead the response from the
+    last registered hook (that is non-null) is used instead.
+
+    Because of parallel processing, this method will be called
+    simultaneously from multiple threads. Make sure your implementation is
+    threadsafe or uses locking.
+
+xml_cache_hit (pkey_object: str)
     This method will be called whenever an XML document is reused from the
     cache. There is one parameter `pkey_object` with the permanent key of
     the DOV object.
@@ -70,7 +117,7 @@ xml_cache_hit
     from multiple threads. Make sure your implementation is threadsafe or uses
     locking.
 
-xml_downloaded
+xml_downloaded (pkey_object: str)
     This method will be called whenever an XML document is downloaded from
     the DOV webservices. There is one parameter `pkey_object` with the
     permanent key of the DOV object.
