+++ pydov/util/hooks.py
@@ -8,7 +8,35 @@ import sys
 import time
 
 
-class AbstractHook(object):
+class Hooks(list):
+    """Runtime representation of registered pydov hooks, i.e. a list of
+    instances of AbstractReadHook and/or AbstractInjectHook."""
+    def get_read_hooks(self):
+        """Get the registered read hooks (i.e. hooks that are subclasses of
+        AbstractReadHook), in the order they are defined in the list.
+
+        Returns
+        -------
+        tuple of AbstractReadHook
+            A tuple with the registered read hooks.
+
+        """
+        return (h for h in self if isinstance(h, AbstractReadHook))
+
+    def get_inject_hooks(self):
+        """Get the registered inject hooks (i.e. hooks that are subclasses of
+        AbstractInjectHook), in the order they are defined in the list.
+
+        Returns
+        -------
+        tuple of AbstractInjectHook
+            A tuple with the registered inject hooks.
+
+        """
+        return (h for h in self if isinstance(h, AbstractInjectHook))
+
+
+class AbstractReadHook(object):
     """Abstract base class for custom hook implementations.
 
     Provides all available methods with a default implementation to do
@@ -34,36 +62,6 @@ class AbstractHook(object):
 
         """
         pass
-
-    def inject_meta_response(self, url):
-        """Inject a response for a metadata request.
-
-        This allows to intercept a metadata request and return a response of
-        your choice.
-
-        When at least one registered hook returns a response for a given URL,
-        the remote call is not executed and instead the response from the
-        last registered hook (that is non-null) is used instead.
-
-        Metadata calls include amongst others: WFS GetCapabilities, requests
-        for MD_Metadata, FC_FeatureCatalogue and XSD schemas.
-
-        These are all calls except for WFS GetFeature requests and XML
-        downloads of DOV data - these are other hooks.
-
-        Parameters
-        ----------
-        url : str
-            URL of the metadata request.
-
-        Returns
-        -------
-        bytes, optional
-            The response to use in favor of resolving the URL. Return None to
-            disable this inject hook.
-
-        """
-        return None
 
     def wfs_search_init(self, typename):
         """Called upon starting a WFS search.
@@ -103,30 +101,6 @@ class AbstractHook(object):
         """
         pass
 
-    def inject_wfs_getfeature_response(self, query):
-        """Inject a response for a WFS GetFeature request.
-
-        This allows to intercept a WFS GetFeature request and return a
-        response of your choice.
-
-        When at least one registered hook returns a response for a given query,
-        the remote call is not executed and instead the response from the
-        last registered hook (that is non-null) is used instead.
-
-        Parameters
-        ----------
-        query : etree.ElementTree
-            The WFS GetFeature request sent to the WFS server.
-
-        Returns
-        -------
-        xml: bytes, optional
-            The GetFeature response to use in favor of resolving the URL.
-            Return None to disable this inject hook.
-
-        """
-        return None
-
     def xml_received(self, pkey_object, xml):
         """Called when the XML of a given object is received, either from
         the cache or from the remote DOV service.
@@ -147,34 +121,6 @@ class AbstractHook(object):
 
         """
         pass
-
-    def inject_xml_response(self, pkey_object):
-        """Inject a response for a DOV XML request.
-
-        This allows to intercept a DOV XML request and return a response of
-        your choice.
-
-        When at least one registered hook returns a response for a given pkey,
-        the remote call is not executed and instead the response from the
-        last registered hook (that is non-null) is used instead.
-
-        Because of parallel processing, this method will be called
-        simultaneously from multiple threads. Make sure your implementation is
-        threadsafe or uses locking.
-
-        Parameters
-        ----------
-        query : etree.ElementTree
-            The WFS GetFeature request sent to the WFS server.
-
[…]
