### 2020-02-07 8b7f283
Centralize inject hook execution.
