### 2020-02-07 89e4e26
Split execution into HookRunner class.
