### 2020-02-07 f0e4f54
Use HookRunner for read hooks too.
