### 2020-02-07 0b672ea
Update docstrings.
