### 2020-02-10 f39ee27
Ignore FutureWarning from OWSLib. (#248)
This used to be here but got accidentally removed in #238.
