### 2020-02-26 c666a6e
Refactor and extend hooks system (#247)
* Add new hooks.

* Add docstrings and remove two hooks.

Remove wfs_search_query and xml_requested hooks.

* More consistent hook naming.

* Update tests.

* Update tests.

* Remove unused import.

* Use UTF8 encoding.

* Update documentation.

* Split read and inject hooks.

* Centralize inject hook execution.

* Split execution into HookRunner class.

* Use HookRunner for read hooks too.

* Update docstrings.

* Update documentation.

* Fix example.

* Remove unused import.

* Fix typo.
