+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 envlist = {py35,py36,py37,py38}-{nolxml,lxml}, update-oefen, update-productie, flake8, docs
 
-[travis]
+[gh-actions]
 python =
     3.8: py38
     3.7: py37
