+++ .github/workflows/ci.yml
@@ -6,6 +6,8 @@ on:
     branches: ["master"]
   pull_request:
     branches: ["master"]
+  schedule:
+    cron: '0 1 1 * *'
 
 jobs:
   tests:
