+++ .github/workflows/ci.yml
@@ -7,7 +7,7 @@ on:
   pull_request:
     branches: ["master"]
   schedule:
-    cron: '0 1 1 * *'
+    - cron: '0 1 1 * *'
 
 jobs:
   tests:
