### 2020-03-16 9db5a89
Apply tests on different os
