+++ .github/workflows/ci.yml
@@ -16,6 +16,7 @@ jobs:
 
     strategy:
       matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: ["3.5", "3.6", "3.7", "3.8"]
 
     steps:
