### 2020-03-16 fd55093
Add python 3.8 test env to gh-actions
