+++ tox.ini
@@ -3,7 +3,7 @@ envlist = {py35,py36,py37,py38}-{nolxml,lxml}, update-oefen, update-productie, f
 
 [gh-actions]
 python =
-    3.8: py38
+    3.8: py38, update-oefen, update-productie, docs
     3.7: py37
     3.6: py36
     3.5: py35
