### 2020-03-16 628bf1a
Improve env description
