### 2020-03-16 7a01515
Continue when single job fails
