+++ .github/workflows/ci.yml
@@ -18,6 +18,8 @@ jobs:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: ["3.5", "3.6", "3.7", "3.8"]
+        fail-fast: 
+          - false
 
     steps:
       - uses: "actions/checkout@v2"
