+++ .github/workflows/ci.yml
@@ -10,16 +10,18 @@ on:
     - cron: '0 1 1 * *'
 
 jobs:
-  tests:
+  tests-linuxmac:
     name: "Python ${{ matrix.python-version }} on ${{ matrix.os }}"
     runs-on: ${{ matrix.os }}
     
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        os: [ubuntu-latest]
         python-version: ["3.5", "3.6", "3.7", "3.8"]
-        fail-fast: 
-          - false
+        fail-fast: [false]
+        include: 
+          - os: macos-latest
+            python-version: 3.8  
 
     steps:
       - uses: "actions/checkout@v2"
@@ -29,6 +31,31 @@ jobs:
       - name: "Install dependencies"
         run: |
           set -xe
+          python -VV
+          python -m site
+          python -m pip install --upgrade pip setuptools wheel
+          python -m pip install --upgrade coverage[toml] virtualenv tox tox-gh-actions
+
+      - name: "Run tox targets for ${{ matrix.python-version }}"
+        run: "python -m tox"
+
+  tests-windows:
+    name: "Python ${{ matrix.python-version }} on ${{ matrix.os }}"
+    runs-on: ${{ matrix.os }}
+    
+    strategy:
+      matrix:
+        os: [windows-latest]
+        python-version: ["3.5", "3.6", "3.7", "3.8"]
+        fail-fast: [false]
+
+    steps:
+      - uses: "actions/checkout@v2"
+      - uses: "actions/setup-python@v1"
+        with:
+          python-version: "${{ matrix.python-version }}"
+      - name: "Install dependencies"
+        run: |
           python -VV
           python -m site
           python -m pip install --upgrade pip setuptools wheel
