+++ .github/workflows/ci.yml
@@ -16,13 +16,15 @@ jobs:
     
     strategy:
       matrix:
-        os: [ubuntu-latest]
+        os: [ubuntu-latest, macos-latest]
         python-version: ["3.5", "3.6", "3.7", "3.8"]
         fail-fast: [false]
-        include: 
+        exclude: 
           - os: macos-latest
-            python-version: 3.8  
-
+            python-version: 3.5
+          - os: macos-latest
+            python-version: 3.6
+          
     steps:
       - uses: "actions/checkout@v2"
       - uses: "actions/setup-python@v1"
