### 2020-03-16 970cf76
Add pandoc for ubuntu
