+++ .github/workflows/ci.yml
@@ -10,27 +10,22 @@ on:
     - cron: '0 1 1 * *'
 
 jobs:
-  tests-linuxmac:
+  tests-linux:
     name: "Python ${{ matrix.python-version }} on ${{ matrix.os }}"
     runs-on: ${{ matrix.os }}
     
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest]
         python-version: ["3.5", "3.6", "3.7", "3.8"]
         fail-fast: [false]
-        exclude: 
-          - os: macos-latest
-            python-version: 3.5
-          - os: macos-latest
-            python-version: 3.6
           
     steps:
       - uses: "actions/checkout@v2"
       - uses: "actions/setup-python@v1"
         with:
           python-version: "${{ matrix.python-version }}"
-      - if: matrix.os == 'ubuntu-latest'
+      - name: "Install pandoc"
         run: sudo apt install pandoc
       - name: "Install dependencies"
         run: |
@@ -58,6 +53,8 @@ jobs:
       - uses: "actions/setup-python@v1"
         with:
           python-version: "${{ matrix.python-version }}"
+      - name: "Install pandoc"
+        run: choco install pandoc
       - name: "Install dependencies"
         run: |
           python -VV
