### 2020-03-16 1af49e1
Install pandoc only for python 3.8
