+++ .github/workflows/ci.yml
@@ -27,7 +27,7 @@ jobs:
           python-version: "${{ matrix.python-version }}"
       - name: "Install pandoc"
         run: sudo apt install pandoc
-        if: matrix.python-version == "3.8"
+        if: matrix.python-version == '3.8'
       - name: "Install dependencies"
         run: |
           set -xe
@@ -56,7 +56,7 @@ jobs:
           python-version: "${{ matrix.python-version }}"
       - name: "Install pandoc"
         run: choco install pandoc
-        if: matrix.python-version == "3.8"
+        if: matrix.python-version == '3.8'
       - name: "Install dependencies"
         run: |
           python -VV
