### 2020-03-16 907015b
Use call to python instead of python3
