+++ tox.ini
@@ -49,7 +49,7 @@ deps =
     lxml
 commands =
     pip install -U pip
-    python3 {toxinidir}/tests/data/update_test_data.py
+    python {toxinidir}/tests/data/update_test_data.py
     py.test --basetemp={envtmpdir} --cov=pydov
 
 [testenv:update-productie]
@@ -61,7 +61,7 @@ deps =
     lxml
 commands =
     pip install -U pip
-    python3 {toxinidir}/tests/data/update_test_data.py
+    python {toxinidir}/tests/data/update_test_data.py
     py.test --basetemp={envtmpdir} --cov=pydov
 
 ; If you want to make tox run the tests with the same versions, create a
