### 2020-04-07 #19 @lwasser
hey @nickledave and @xmnlab i know there is a lot going on. i just wanted to check in on this review to see how you guys were doing.  I am very behind and know that ropensci is stopping new submissions for atleast a month. are you guys still able to review this package or should we take a step back?

Many thanks.
