### 2020-04-16 #19 @xmnlab
@Roel, @stijnvanhoey, @pjhaest, thanks for your patience. pydov looks awesome, I am adding my review notes below:

- [x] **Function Documentation:** for all user-facing functions
- [x] **Examples** for all user-facing functions
- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, the badge for pyOpenSci peer-review once it has started (see below), a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges, see this example, that one and that one. Such a table should be more wide than high.
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [x] Any additional setup required (authentication tokens, etc)
- [x] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [x] If applicable, how the package compares to other similar packages and/or how it relates to other packages

- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

#### For packages co-submitting to JOSS

---

Congratulation for the work you all are doing on pydov. I am adding general notes below:

- The documentation looks very detailed and instructive about how to use the library. I agree with David, maybe would be good to have a tutorial with an introduction something like a blog post (also it could be posted on PyOpenSci blog section).
- API reference documentation looks very well detailed.
- Good test coverage (96%).

Requests

  - [ ] There is no instructions about how to install the package for development in README file (e.g. prepare the environment: "pip install -r requirements_dev.txt", install the package in the development mode "python setup.py develop", etc)
  - [ ] The quick start located at README maybe could be improved with the output of dataframe, so it will give more details about the result of a boring search. It also applies to https://pydov.readthedocs.io/en/stable/quickstart.html

Suggestions and general comments (not required)

  - Maybe it would be good to have a link to binder on each tutorial page to its corresponding tutorial on binder
  - In the tests, you can change `assert type(df) is DataFrame` to `assert isinstance(df, DataFrame)`
  - It is not too common to import objects from a test_*.py maybe it could be moved to conftest (if it applies)
  - It seems the test classes are inheriting from a base class (AbstractTestTypes, AbstractTestSearch) and some methods' docstring are the same as the inherited method, using this approach is very hard to maintain and keep everything consistent. Also some methods could be an attribute, e.g.: `def get_type(self): return Boring` -> `klass = Boring` and it keeps your test classes more compact.
  - The pydov classes are also inheriting from base classes, and the docstrings work looks duplicated, if you don't add any docstring, the method will inherit from the base clases. If you think you need to have a customized docstring, maybe you can check some approaches for dynamic docstrings
  - typo: https://pydov.readthedocs.io/en/stable/notebooks/search_grondwaterfilters.html#Plotting "... Pandas cen be used"
  - It would be good to have in each chart the title and label for each axis
  - Maybe pydocstyle can help to check static docstrings (you also can ignore erros you think any one doesn't apply for pydov). If you want to check dynamic docstrings you can use numpydoc --validate (https://numpydoc.readthedocs.io/en/latest/validation.html#validating-numpydoc-docstrings) you can check one example here https://github.com/pandas-dev/pandas/blob/master/scripts/validate_docstrings.py
