### 2020-04-18 DOV-Vlaanderen/pydov#255 @Roel
Update development installation instructions
From the [pyOpenSci review](https://github.com/pyOpenSci/software-review/issues/19):

> There is no instructions about how to install the package for development in README file (e.g. prepare the environment: "pip install -r requirements_dev.txt", install the package in the development mode "python setup.py develop", etc)

There are bits and pieces of this information in the docs (in the [installation section](https://pydov.readthedocs.io/en/latest/installation.html#from-sources) and the [contributing section](https://pydov.readthedocs.io/en/latest/contributing.html#setting-up-your-environment) for example), but this might be hard to find and assemble.

I suggest:
- adding development installation instructions to the README
- adding the same instructions in the installation section in the docs
- add cross-reference between installation section and contributing section, maybe with a bit more info in the latter (regarding the different requirements files etc.)
