### 2020-04-18 DOV-Vlaanderen/pydov#256 @Roel
Add dataframe output to README
From the [pyOpenSci review](https://github.com/pyOpenSci/software-review/issues/19):

> The quick start located at README maybe could be improved with the output of dataframe, so it will give more details about the result of a boring search. It also applies to https://pydov.readthedocs.io/en/stable/quickstart.html

We should add the output of the search in the Quick start to the documentation so people know what to expect as a result.

This should be added in the README as well as in the Quick start section in the documentation.
