### 2020-04-18 DOV-Vlaanderen/pydov#257 @Roel
Add Binder link to top of each tutorial notebook
From the [pyOpenSci review](https://github.com/pyOpenSci/software-review/issues/19):

> Maybe it would be good to have a link to binder on each tutorial page to its corresponding tutorial on binder

Next to the global Binder link on the overview Tutorials page, it might be handy to add a link to each tutorial to open an interactive version of that tutorial in Binder.
