### 2020-04-18 DOV-Vlaanderen/pydov#258 @Roel
Typo in tutorial
From the [pyOpenSci review](https://github.com/pyOpenSci/software-review/issues/19):
> typo: https://pydov.readthedocs.io/en/stable/notebooks/search_grondwaterfilters.html#Plotting "... Pandas cen be used"
