### 2020-04-18 DOV-Vlaanderen/pydov#259 @Roel
Add titles and axis labels to all charts
From the [pyOpenSci review](https://github.com/pyOpenSci/software-review/issues/19):
> It would be good to have in each chart the title and label for each axis
