### 2020-04-18 DOV-Vlaanderen/pydov#260 @Roel
Tests: check isinstance instead of type
From the [pyOpenSci review](https://github.com/pyOpenSci/software-review/issues/19):

> In the tests, you can change assert type(df) is DataFrame to assert isinstance(df, DataFrame)

This applies generally I think and needs to be checked and refactored everywhere in the tests (i.e. not only for the DataFrame case)
