### 2020-04-18 DOV-Vlaanderen/pydov#261 @Roel
Tests: move fixtures to conftest.py so they can be used without import
From the [pyOpenSci review](https://github.com/pyOpenSci/software-review/issues/19):
> It is not too common to import objects from a test_*.py maybe it could be moved to conftest (if it applies)

We should check if we can move all pytest fixtures to conftest.py so they can be used without import.
