### 2020-04-18 DOV-Vlaanderen/pydov#262 @Roel
Remove identical docstrings in subclasses
From the [pyOpenSci review](https://github.com/pyOpenSci/software-review/issues/19):

> It seems the test classes are inheriting from a base class (AbstractTestTypes, AbstractTestSearch) and some methods' docstring are the same as the inherited method, using this approach is very hard to maintain and keep everything consistent. Also some methods could be an attribute, e.g.: def get_type(self): return Boring -> klass = Boring and it keeps your test classes more compact.

> The pydov classes are also inheriting from base classes, and the docstrings work looks duplicated, if you don't add any docstring, the method will inherit from the base clases. If you think you need to have a customized docstring, maybe you can check some approaches for dynamic docstrings

I think it is sufficient to keep the docstrings in the base classes only, and only keep them in the subclass when they are different from those of the base class?

See also #263
