### 2020-05-24 #19 @Roel
Hi @lwasser ! Our code sprint is planned for next week (May 28-29), so I hope to make some progress in the issues then. I assume we should be able to pick up the review somewhere in June :)

Thanks for your patience, enjoy your week offline and stay healthy!
