+++ tests/conftest.py
@@ -1,10 +1,16 @@
 """Module grouping session scoped PyTest fixtures."""
 
+import owslib
 import pytest
 from _pytest.monkeypatch import MonkeyPatch
+from owslib.etree import etree
+from owslib.util import findall
+from owslib.wfs import WebFeatureService
 
 import pydov
 from pydov import Hooks
+from pydov.util import owsutil
+from pydov.util.dovutil import build_dov_url
 
 
 def pytest_runtest_setup():
@@ -21,3 +27,365 @@ def monkeymodule():
     mpatch = MonkeyPatch()
     yield mpatch
     mpatch.undo()
+
+
+@pytest.fixture(scope='module')
+def mp_wfs(monkeymodule):
+    """Monkeypatch the call to the remote GetCapabilities request.
+
+    Parameters
+    ----------
+    monkeymodule : pytest.fixture
+        PyTest monkeypatch fixture with module scope.
+
+    """
+    def read(*args, **kwargs):
+        with open('tests/data/util/owsutil/wfscapabilities.xml', 'r',
+                  encoding='utf-8') as f:
+            data = f.read()
+            if type(data) is not bytes:
+                data = data.encode('utf-8')
+            data = etree.fromstring(data)
+        return data
+
+    monkeymodule.setattr(
+        owslib.feature.common.WFSCapabilitiesReader, 'read', read)
+
+
+@pytest.fixture(scope='module')
+def wfs(mp_wfs):
+    """PyTest fixture providing an instance of a WebFeatureService based on
+    a local copy of a GetCapabilities request.
+
+    Parameters
+    ----------
+    mp_wfs : pytest.fixture
+        Monkeypatch the call to the remote GetCapabilities request.
+
+    Returns
+    -------
+    owslib.wfs.WebFeatureService
+        WebFeatureService based on the local GetCapabilities.
+
+    """
+    return WebFeatureService(
+        url=build_dov_url('geoserver/wfs'), version="1.1.0")
+
+
+@pytest.fixture()
+def mp_remote_fc_notfound(monkeypatch):
+    """Monkeypatch the call to get an inexistent remote featurecatalogue.
+
+    Parameters
+    ----------
+    monkeypatch : pytest.fixture
+        PyTest monkeypatch fixture.
+
+    """
+    def __get_remote_fc(*args, **kwargs):
+        with open('tests/data/util/owsutil/fc_featurecatalogue_notfound.xml',
+                  'r') as f:
+            data = f.read()
+            if type(data) is not bytes:
+                data = data.encode('utf-8')
+        return data
+
+    monkeypatch.setattr(pydov.util.owsutil, '__get_remote_fc', __get_remote_fc)
+
+
+@pytest.fixture(scope='module')
+def mp_remote_md(wfs, monkeymodule, request):
+    """Monkeypatch the call to get the remote metadata of the layer.
+
+    This monkeypatch requires a module variable ``location_md_metadata``
+    with the path to the md_metadata file on disk.
+
+    Parameters
+    ----------
+    wfs : pytest.fixture returning owslib.wfs.WebFeatureService
+        WebFeatureService based on the local GetCapabilities.
+    monkeymodule : pytest.fixture
+        PyTest monkeypatch fixture with module scope.
+    request : pytest.fixtue
+        PyTest fixture providing request context.
+
+    """
+    def __get_remote_md(*args, **kwargs):
+        file_path = getattr(request.module, "location_md_metadata")
+        with open(file_path, 'r') as f:
+            data = f.read()
+            if type(data) is not bytes:
+                data = data.encode('utf-8')
+        return MD_Metadata(etree.fromstring(data).find(
[…]
