+++ tests/test_types_itp_informelehydrogeologischestratigrafie.py
@@ -1,20 +1,12 @@
 """Module grouping tests for the
 pydov.types.interpretaties.FormeleStratigrafie class."""
-from pydov.types.interpretaties import (
-    FormeleStratigrafie,
-    InformeleHydrogeologischeStratigrafie,
-)
+from pydov.types.interpretaties import (FormeleStratigrafie,
+                                        InformeleHydrogeologischeStratigrafie)
 from pydov.util.dovutil import build_dov_url
 from tests.abstract import AbstractTestTypes
-
-from tests.test_search_itp_formelestratigrafie import (
-    wfs_feature,
-    wfs_getfeature,
-    mp_dov_xml,
-    location_wfs_feature,
-    location_wfs_getfeature,
-    location_dov_xml,
-)
+from tests.test_search_itp_formelestratigrafie import (location_dov_xml,
+                                                       location_wfs_feature,
+                                                       location_wfs_getfeature)
 
 
 class TestInformeleHydrogeologischeFormeleStratigrafie(AbstractTestTypes):
