+++ tests/abstract.py
@@ -1047,7 +1047,7 @@ class AbstractTestTypes(object):
 
         Parameters
         ----------
-        wfs_getfeature : pytest.fixture returing str
+        wfs_getfeature : pytest.fixture returning str
             Fixture providing a WFS GetFeature response.
 
         """
@@ -1063,7 +1063,7 @@ class AbstractTestTypes(object):
 
         Parameters
         ----------
-        wfs_getfeature : pytest.fixture returing str
+        wfs_getfeature : pytest.fixture returning str
             Fixture providing a WFS GetFeature response.
 
         """
@@ -1079,7 +1079,7 @@ class AbstractTestTypes(object):
 
         Parameters
         ----------
-        wfs_getfeature : pytest.fixture returing str
+        wfs_getfeature : pytest.fixture returning str
             Fixture providing a WFS GetFeature response.
 
         """
@@ -1095,7 +1095,7 @@ class AbstractTestTypes(object):
 
         Parameters
         ----------
-        wfs_getfeature : pytest.fixture returing str
+        wfs_getfeature : pytest.fixture returning str
             Fixture providing a WFS GetFeature response.
 
         """
