+++ tests/test_util_caching.py
@@ -46,7 +46,7 @@ def plaintext_cache(request):
     This fixture should be parametrized, with a list of parameters in the
     order described below.
 
-    Paramaters
+    Parameters
     ----------
     max_age : datetime.timedelta
         The maximum age to use for the cache.
@@ -77,7 +77,7 @@ def gziptext_cache(request):
     This fixture should be parametrized, with a list of parameters in the
     order described below.
 
-    Paramaters
+    Parameters
     ----------
     max_age : datetime.timedelta
         The maximum age to use for the cache.
@@ -361,7 +361,7 @@ class TestPlainTextFileCacheCache(object):
     def test_return_type(self, plaintext_cache, mp_remote_xml):
         """The the return type of the get method.
 
-        Test wether the get method returns the data in the same datatype (
+        Test whether the get method returns the data in the same datatype (
         i.e. bytes) regardless of the data was cached or not.
 
         Parameters
@@ -643,7 +643,7 @@ class TestGzipTextFileCacheCache(object):
     def test_return_type(self, gziptext_cache, mp_remote_xml):
         """The the return type of the get method.
 
-        Test wether the get method returns the data in the same datatype (
+        Test whether the get method returns the data in the same datatype (
         i.e. bytes) regardless of the data was cached or not.
 
         Parameters
