+++ docs/index.rst
@@ -29,7 +29,7 @@ aggregates data about soil, subsoil and groundwater of Flanders and makes them p
 available. Interactive and human-readable extraction and querying of the data is provided
 by a `web application <https://www.dov.vlaanderen.be/portaal/?module=verkenner>`_\ , whereas
 the focus of this package is to support **machine-based** extraction and conversion of
-the data. The pacakge aims to support a set of complementary use cases, for example:
+the data. The package aims to support a set of complementary use cases, for example:
 
 * integrate DOV data in larger data processing pipelines
 * support the reproducibility and/or repeatability of research studies
@@ -54,7 +54,7 @@ for each of the object types (:class:`~pydov.search.grondwaterfilter.GrondwaterF
 :class:`~pydov.search.interpretaties.InformeleStratigrafieSearch`).
 
 The setup of appropriate queries on attribute or location is similar for the different data types and
-crucial to get acces to the required data. The :ref:`query_attribute` and :ref:`query_location` explain
+crucial to get access to the required data. The :ref:`query_attribute` and :ref:`query_location` explain
 the query options for attributes and locations respectively.
 
 All functionalities are build on top of the existing webservices provided
