+++ tests/test_search_itp_quartairstratigrafie.py
@@ -1,11 +1,7 @@
 """Module grouping tests for the interpretaties search module."""
-import numpy as np
-import pandas as pd
-import pytest
 from owslib.fes import PropertyIsEqualTo
 from pandas import DataFrame
 
-import pydov
 from pydov.search.interpretaties import QuartairStratigrafieSearch
 from pydov.types.interpretaties import QuartairStratigrafie
 from tests.abstract import AbstractTestSearch
@@ -13,15 +9,18 @@ from tests.abstract import AbstractTestSearch
 location_md_metadata = \
     'tests/data/types/interpretaties/quartaire_stratigrafie/md_metadata.xml'
 location_fc_featurecatalogue = \
-    'tests/data/types/interpretaties/quartaire_stratigrafie/fc_featurecatalogue.xml'
+    'tests/data/types/interpretaties/quartaire_stratigrafie/' \
+    'fc_featurecatalogue.xml'
 location_wfs_describefeaturetype = \
-    'tests/data/types/interpretaties/quartaire_stratigrafie/wfsdescribefeaturetype.xml'
+    'tests/data/types/interpretaties/quartaire_stratigrafie/' \
+    'wfsdescribefeaturetype.xml'
 location_wfs_getfeature = \
     'tests/data/types/interpretaties/quartaire_stratigrafie/wfsgetfeature.xml'
 location_wfs_feature = \
     'tests/data/types/interpretaties/quartaire_stratigrafie/feature.xml'
 location_dov_xml = \
-    'tests/data/types/interpretaties/quartaire_stratigrafie/quartaire_stratigrafie.xml'
+    'tests/data/types/interpretaties/quartaire_stratigrafie/' \
+    'quartaire_stratigrafie.xml'
 location_xsd_base = \
     'tests/data/types/interpretaties/quartaire_stratigrafie/xsd_*.xml'
 
