+++ tests/test_util_query.py
@@ -1,13 +1,10 @@
 """Module grouping tests for the pydov.util.query module."""
-import pandas as pd
 import numpy as np
+import pandas as pd
 import pytest
 
 from pydov.util.dovutil import build_dov_url
-from pydov.util.query import (
-    PropertyInList,
-    Join,
-)
+from pydov.util.query import Join, PropertyInList
 
 
 class TestPropertyInList(object):
