### 2020-05-28 1d44385
Avoid cross-importing testdata location in types tests.
