+++ tests/test_types_grondwaterfilter.py
@@ -2,9 +2,10 @@
 from pydov.types.grondwaterfilter import GrondwaterFilter
 from pydov.util.dovutil import build_dov_url
 from tests.abstract import AbstractTestTypes
-from tests.test_search_grondwaterfilter import (location_dov_xml,
-                                                location_wfs_feature,
-                                                location_wfs_getfeature)
+
+location_wfs_getfeature = 'tests/data/types/grondwaterfilter/wfsgetfeature.xml'
+location_wfs_feature = 'tests/data/types/grondwaterfilter/feature.xml'
+location_dov_xml = 'tests/data/types/grondwaterfilter/grondwaterfilter.xml'
 
 
 class TestGrondwaterFilter(AbstractTestTypes):
