+++ tests/test_types_itp_quartairstratigrafie.py
@@ -3,8 +3,14 @@ pydov.types.interpretaties.QuartairStratigrafie class."""
 from pydov.types.interpretaties import QuartairStratigrafie
 from pydov.util.dovutil import build_dov_url
 from tests.abstract import AbstractTestTypes
-from tests.test_search_itp_quartairstratigrafie import (
-    location_dov_xml, location_wfs_feature, location_wfs_getfeature)
+
+location_wfs_getfeature = \
+    'tests/data/types/interpretaties/quartaire_stratigrafie/wfsgetfeature.xml'
+location_wfs_feature = \
+    'tests/data/types/interpretaties/quartaire_stratigrafie/feature.xml'
+location_dov_xml = \
+    'tests/data/types/interpretaties/quartaire_stratigrafie/' \
+    'quartaire_stratigrafie.xml'
 
 
 class TestQuartairStratigrafie(AbstractTestTypes):
