+++ tests/test_types_itp_lithologischebeschrijvingen.py
@@ -3,8 +3,15 @@ pydov.types.interpretaties.LithologischeBeschrijvingen class."""
 from pydov.types.interpretaties import LithologischeBeschrijvingen
 from pydov.util.dovutil import build_dov_url
 from tests.abstract import AbstractTestTypes
-from tests.test_search_itp_lithologischebeschrijvingen import (
-    location_dov_xml, location_wfs_feature, location_wfs_getfeature)
+
+location_wfs_getfeature = \
+    'tests/data/types/interpretaties/lithologische_beschrijvingen/' \
+    'wfsgetfeature.xml'
+location_wfs_feature = \
+    'tests/data/types/interpretaties/lithologische_beschrijvingen/feature.xml'
+location_dov_xml = \
+    'tests/data/types/interpretaties/lithologische_beschrijvingen' \
+    '/lithologische_beschrijvingen.xml'
 
 
 class TestLithologischeBeschrijvingen(AbstractTestTypes):
