+++ tests/test_types_itp_formelestratigrafie.py
@@ -3,9 +3,14 @@ pydov.types.interpretaties.FormeleStratigrafie class."""
 from pydov.types.interpretaties import FormeleStratigrafie
 from pydov.util.dovutil import build_dov_url
 from tests.abstract import AbstractTestTypes
-from tests.test_search_itp_formelestratigrafie import (location_dov_xml,
-                                                       location_wfs_feature,
-                                                       location_wfs_getfeature)
+
+location_wfs_getfeature = \
+    'tests/data/types/interpretaties/formele_stratigrafie/wfsgetfeature.xml'
+location_wfs_feature = \
+    'tests/data/types/interpretaties/formele_stratigrafie/feature.xml'
+location_dov_xml = \
+    'tests/data/types/interpretaties/formele_stratigrafie' \
+    '/formele_stratigrafie.xml'
 
 
 class TestFormeleStratigrafie(AbstractTestTypes):
