### 2020-05-28 02e7b83
Change all "assert type(xxx) is yyy" to "assert isinstance(xxx,yyy)" in tests.
