+++ tests/abstract.py
@@ -243,34 +243,34 @@ class AbstractTestSearch(object):
         """
         fields = self.get_search_object().get_fields()
 
-        assert type(fields) is dict
+        assert isinstance(fields, dict)
 
         for field in fields:
-            assert type(field) is str
+            assert isinstance(field, str)
 
             f = fields[field]
-            assert type(f) is dict
+            assert isinstance(f, dict)
 
             assert 'name' in f
-            assert type(f['name']) is str
+            assert isinstance(f['name'],str)
             assert f['name'] == field
 
             assert 'definition' in f
-            assert type(f['name']) is str
+            assert isinstance(f['name'], str)
 
             assert 'type' in f
-            assert type(f['type']) is str
+            assert isinstance(f['type'], str)
             assert f['type'] in ['string', 'float', 'integer', 'date',
                                  'datetime', 'boolean']
 
             assert 'notnull' in f
-            assert type(f['notnull']) is bool
+            assert isinstance(f['notnull'], bool)
 
             assert 'query' in f
-            assert type(f['query']) is bool
+            assert isinstance(f['query'], bool)
 
             assert 'cost' in f
-            assert type(f['cost']) is int
+            assert isinstance(f['cost'], bool)
             assert f['cost'] > 0
 
             if 'values' in f:
@@ -278,22 +278,22 @@ class AbstractTestSearch(object):
                     'cost', 'definition', 'name', 'notnull', 'query', 'type',
                     'values']
 
-                assert type(f['values']) is dict
+                assert isinstance(f['values'], dict)
 
                 for v in f['values'].keys():
-                    assert type(f['values'][v]) is str or f[
+                    assert isinstance(f['values'][v], str) or f[
                         'values'][v] is None
 
                     if f['type'] == 'string':
-                        assert type(v) is str
+                        assert isinstance(v, str)
                     elif f['type'] == 'float':
-                        assert type(v) is float
+                        assert isinstance(v, float)
                     elif f['type'] == 'integer':
-                        assert type(v) is int
+                        assert isinstance(v, int)
                     elif f['type'] == 'date':
-                        assert type(v) is datetime.date
+                        assert isinstance(v, datetime.date)
                     elif f['type'] == 'boolean':
-                        assert type(v) is bool
+                        assert isinstance(v, bool)
             else:
                 assert sorted(f.keys()) == ['cost', 'definition', 'name',
                                             'notnull', 'query', 'type']
@@ -320,7 +320,7 @@ class AbstractTestSearch(object):
             query=self.get_valid_query_single(),
             return_fields=self.get_valid_returnfields())
 
-        assert type(df) is DataFrame
+        assert isinstance(df, DataFrame)
 
     def test_search(self, mp_wfs, mp_get_schema, mp_remote_describefeaturetype,
                     mp_remote_md, mp_remote_fc, mp_remote_wfs_feature,
@@ -350,7 +350,7 @@ class AbstractTestSearch(object):
         df = self.get_search_object().search(
             query=self.get_valid_query_single())
 
-        assert type(df) is DataFrame
+        assert isinstance(df, DataFrame)
 
         assert list(df) == self.get_df_default_columns()
 
@@ -410,7 +410,7 @@ class AbstractTestSearch(object):
             query=self.get_valid_query_single(),
             return_fields=self.get_valid_returnfields())
 
-        assert type(df) is DataFrame
+        assert isinstance(df, DataFrame)
 
         assert list(df) == list(self.get_valid_returnfields())
 
@@ -431,7 +431,7 @@ class AbstractTestSearch(object):
             query=self.get_valid_query_single(),
[…]
