+++ tests/test_search_itp_informelestratigrafie.py
@@ -218,7 +218,7 @@ class TestInformeleStratigrafieSearch(AbstractTestSearch):
             return_fields=('pkey_interpretatie', 'pkey_boring',
                            'pkey_sondering'))
 
-        assert type(df) is DataFrame
+        assert isinstance(df, DataFrame)
 
         assert list(df) == ['pkey_interpretatie', 'pkey_boring',
                             'pkey_sondering']
