+++ tests/test_search_itp_informelehydrogeologischestratigrafie.py
@@ -201,7 +201,7 @@ class TestInformeleHydrogeologischeStratigrafieSearch(AbstractTestSearch):
             return_fields=('pkey_interpretatie', 'pkey_boring',
                            'gemeente'))
 
-        assert type(df) is DataFrame
+        assert isinstance(df, DataFrame)
 
         assert list(df) == ['pkey_interpretatie', 'pkey_boring',
                             'gemeente']
