+++ tests/test_search_itp_formelestratigrafie.py
@@ -218,7 +218,7 @@ class TestFormeleStratigrafieSearch(AbstractTestSearch):
             return_fields=('pkey_interpretatie', 'pkey_boring',
                            'pkey_sondering'))
 
-        assert type(df) is DataFrame
+        assert isinstance(df, DataFrame)
 
         assert list(df) == ['pkey_interpretatie', 'pkey_boring',
                             'pkey_sondering']
