### 2020-05-28 0cf7f0b
Remove duplicated docstrings from code.
