+++ pydov/search/boring.py
@@ -3,9 +3,10 @@
 import pandas as pd
 
 from pydov.types.fields import _WfsInjectedField
-from .abstract import AbstractSearch
+
 from ..types.boring import Boring
 from ..util import owsutil
+from .abstract import AbstractSearch
 
 
 class BoringSearch(AbstractSearch):
@@ -31,13 +32,10 @@ class BoringSearch(AbstractSearch):
         super(BoringSearch, self).__init__('dov-pub:Boringen', objecttype)
 
     def _init_namespace(self):
-        """Initialise the WFS namespace associated with the layer."""
         if BoringSearch.__wfs_namespace is None:
             BoringSearch.__wfs_namespace = self._get_namespace()
 
     def _init_fields(self):
-        """Initialise the fields and their metadata available in this search
-        class."""
         if self._fields is None:
             if BoringSearch.__wfs_schema is None:
                 BoringSearch.__wfs_schema = self._get_schema()
@@ -76,60 +74,6 @@ class BoringSearch(AbstractSearch):
 
     def search(self, location=None, query=None,
                sort_by=None, return_fields=None, max_features=None):
-        """Search for boreholes (Boring). Provide `location` and/or `query`
-        and/or `max_features`.
-        When `return_fields` is None, all fields are returned.
-
-        Parameters
-        ----------
-        location : pydov.util.location.AbstractLocationFilter or \
-                    owslib.fes.BinaryLogicOpType<AbstractLocationFilter> or \
-                    owslib.fes.UnaryLogicOpType<AbstractLocationFilter>
-            Location filter limiting the features to retrieve. Can either be a
-            single instance of a subclass of AbstractLocationFilter, or a
-            combination using And, Or, Not of AbstractLocationFilters.
-        query : owslib.fes.OgcExpression
-            OGC filter expression to use for searching. This can contain any
-            combination of filter elements defined in owslib.fes. The query
-            should use the fields provided in `get_fields()`. Note that not
-            all fields are currently supported as a search parameter.
-        sort_by : owslib.fes.SortBy, optional
-            List of properties to sort by.
-        return_fields : list<str> or tuple<str> or set<str>
-            A list of fields to be returned in the output data. This should
-            be a subset of the fields provided in `get_fields()`. Note that
-            not all fields are currently supported as return fields.
-        max_features : int
-            Limit the maximum number of features to request.
-
-        Returns
-        -------
-        pandas.core.frame.DataFrame
-            DataFrame containing the output of the search query.
-
-        Raises
-        ------
-        pydov.util.errors.InvalidSearchParameterError
-            When not one of `location`, `query` or `max_features` is provided.
-
-        pydov.util.errors.InvalidFieldError
-            When at least one of the fields in `return_fields` is unknown.
-
-            When a field that is only accessible as return field is used as
-            a query parameter.
-
-            When a field that can only be used as a query parameter is used as
-            a return field.
-
-        pydov.util.errors.FeatureOverflowError
-            When the number of features to be returned is equal to the
-            maxFeatures limit of the WFS server.
-
-        AttributeError
-            When the argument supplied as return_fields is not a list,
-            tuple or set.
-
-        """
         fts = self._search(location=location, query=query, sort_by=sort_by,
                            return_fields=return_fields,
                            max_features=max_features)
