### 2020-05-28 DOV-Vlaanderen/pydov#255 @stijnvanhoey
Basically, the dev get started boils down to...

### Clone repo

```
git clone git@github.com:DOV-Vlaanderen/pydov.git
```

### a development environment

```
conda create -n pydov python=3.7
conda activate pydov
```

or

```
virtualenv pydov
source pydov/bin/activate | pydov\Scripts\activate
```

### Install dependencies for development

Inside the environment:

```
pip install -e .[devs]
```

I would put this in the README and docs installation section and refer to the contirbuting guidelines for more in-depth information. I'm rewriting these guidelines as well to better align.
