### 2020-05-28 e9d7299
Remove duplicated docstrings from tests.
