+++ docs/quickstart.rst
@@ -29,14 +29,14 @@ Currently, we support the following datasets:
 
 Each of the datasets can be queried using a search object for this dataset. While the search objects are different, the workflow is the same for each dataset. Relevant classes can be imported from the pydov.search package, for example if we'd like to query the boreholes dataset:
 
-::
+.. code-block:: python
 
     from pydov.search.boring import BoringSearch
     boringsearch = BoringSearch()
 
 Now we can query for boreholes either on attributes, on location or on a combination of both. To query on attributes, we use the OGC filter functions from OWSLib. For example, to request all boreholes with a depth over 550 m, we would use the following ``query`` parameter:
 
-::
+.. code-block:: python
 
     from owslib.fes import PropertyIsGreaterThan
 
@@ -44,30 +44,90 @@ Now we can query for boreholes either on attributes, on location or on a combina
         query=PropertyIsGreaterThan(
             propertyname='diepte_tot_m', literal='550')
     )
+    dataframe
+
+pydov will load (status is shown in the progress bar) the required data and return the data into a Pandas DataFrame:
+
+::
+
+    [000/253] ..................................................
+    [050/253] ..................................................
+    [100/253] ..................................................
+    [150/253] ..................................................
+    [200/253] ..................................................
+    [250/253] ...
+
+                                               pkey_boring     boornummer         x         y  mv_mtaw  start_boring_mtaw           gemeente  diepte_boring_van  diepte_boring_tot datum_aanvang                          uitvoerder  boorgatmeting  diepte_methode_van  diepte_methode_tot  boormethode
+    0    https://www.dov.vlaanderen.be/data/boring/1965...  kb15d27e-B149  144820.0  217840.0     2.00               2.00            Beveren                0.0             622.00    1965-07-13  Belgische Geologische Dienst (BGD)           True                0.00              622.00     onbekend
+    1    https://www.dov.vlaanderen.be/data/boring/1984...    kb9d9w-B244  200063.0  235530.0    30.44              30.44             Ravels                0.0             570.00    1984-03-19                       Smet - Dessel           True                0.00              570.00  spoelboring
+    2    https://www.dov.vlaanderen.be/data/boring/2016...  kb25d61e-B348  209825.6  195829.2    38.00              38.00           Beringen                0.0             575.00           NaN                            onbekend          False                0.00              575.00     onbekend
+    3    https://www.dov.vlaanderen.be/data/boring/1901...    kb26d63e-B1  237924.0  194897.0    65.00              65.00         Opglabbeek                0.0             713.27    1901-01-01  Belgische Geologische Dienst (BGD)          False                0.00              713.27     onbekend
+    4    https://www.dov.vlaanderen.be/data/boring/1902...    kb26d63e-B2  235262.0  197398.0    75.30              75.30  Meeuwen-Gruitrode                0.0             870.10    1902-01-01  Belgische Geologische Dienst (BGD)          False                0.00              870.10     onbekend
+    ..                                                 ...            ...       ...       ...      ...                ...                ...                ...                ...           ...                                 ...            ...                 ...                 ...          ...
+    405  https://www.dov.vlaanderen.be/data/boring/2002...    BGD048e0294  238725.0  201000.0    51.00              51.00               Bree                0.0             571.15    2002-05-21                          Smet - GWT           True              416.09              566.14    rollerbit
[…]
