### 2020-05-28 f0b8edc
Extend installation instructions
