+++ README.md
@@ -10,13 +10,25 @@ Please note that downloading DOV data with pydov is governed by the same [discla
 
 ## Installation
 
-You can install `pydov` stable easily using pip:
+You can install `pydov` stable using pip:
 
 ```shell script
 pip install pydov
 ```
 
-Or clone the Git repository and install with `setup.py` to get the latest snapshot from the master branch.
+Or clone the [git repository](https://github.com/DOV-Vlaanderen/pydov) and install with `python setup.py install` to get the latest snapshot from the master branch.
+
+To contribute to the code, make sure to install the package and all of the development dependencies enlisted in the
+[requiremends_dev.txt](requiremends_dev.txt) file. First, clone the [git repository](https://github.com/DOV-Vlaanderen/pydov).
+We advice to use an Python development environment, for example with [conda](https://docs.conda.io/en/latest/miniconda.html) or
+[virtualenv](https://virtualenv.pypa.io/en/latest/). Activate the (conda/virtualenv) environment and
+install the package in development mode:
+
+```shell script
+pip install -e .[devs]
+```
+
+Need more detailed instructions? Check out the [installation instructions](https://pydov.readthedocs.io/en/stable/installation.html) and the [contributing guidelines](https://pydov.readthedocs.io/en/stable/contributing.html#development-guidelines).
 
 ## Quick start
 
@@ -45,9 +57,5 @@ Full documentation of `pydov` can be found on our [ReadTheDocs page](https://pyd
 - We welcome [contributions](.github/CONTRIBUTING.rst) including bug reports.
 - License: MIT
 - Citation information can be found on [Zenodo](https://doi.org/10.5281/zenodo.2788680).
-
-
 - Please note that this project is released with a [Contributor Code of Conduct](.github/CODE_OF_CONDUCT.rst). By participating in this project you agree to abide by its terms.
-
-
 - Also note that downloading DOV data with pydov is governed by the same [disclaimer](https://www.dov.vlaanderen.be/page/disclaimer) that applies to the other DOV services. Be sure to consult it when using DOV data with pydov.
