### 2020-05-28 26fef61
Refactor AbstractTestSearch to use class variables instead of getters.
