+++ tests/test_search_itp_gecodeerdelithologie.py
@@ -37,44 +37,32 @@ location_xsd_base = \
 
 
 class TestGecodeerdeLithologieSearch(AbstractTestSearch):
-    def get_search_object(self):
-        return GecodeerdeLithologieSearch()
 
-    def get_type(self):
-        return GecodeerdeLithologie
+    search_instance = GecodeerdeLithologieSearch()
+    datatype_class = GecodeerdeLithologie
 
-    def get_valid_query_single(self):
-        return PropertyIsEqualTo(propertyname='Proefnummer',
-                                 literal='kb22d56w-B165')
+    valid_query_single = PropertyIsEqualTo(propertyname='Proefnummer',
+                                           literal='kb22d56w-B165')
 
-    def get_inexistent_field(self):
-        return 'onbestaand'
+    inexistent_field = 'onbestaand'
+    wfs_field = 'Proefnummer'
+    xml_field = 'grondsoort'
 
-    def get_xml_field(self):
-        return 'grondsoort'
+    valid_returnfields = ('pkey_interpretatie',
+                          'betrouwbaarheid_interpretatie')
+    valid_returnfields_subtype = (
+        'pkey_interpretatie', 'diepte_laag_van', 'diepte_laag_tot')
+    valid_returnfields_extra = ('pkey_interpretatie', 'gemeente')
 
-    def get_wfs_field(self):
-        return 'Proefnummer'
-
-    def get_valid_returnfields(self):
-        return ('pkey_interpretatie', 'betrouwbaarheid_interpretatie')
-
-    def get_valid_returnfields_subtype(self):
-        return ('pkey_interpretatie', 'diepte_laag_van', 'diepte_laag_tot')
-
-    def get_valid_returnfields_extra(self):
-        return ('pkey_interpretatie', 'gemeente')
-
-    def get_df_default_columns(self):
-        return ['pkey_interpretatie', 'pkey_boring',
-                'betrouwbaarheid_interpretatie', 'x', 'y',
-                'diepte_laag_van', 'diepte_laag_tot',
-                'hoofdnaam1_grondsoort', 'hoofdnaam2_grondsoort',
-                'bijmenging1_plaatselijk', 'bijmenging1_hoeveelheid',
-                'bijmenging1_grondsoort', 'bijmenging2_plaatselijk',
-                'bijmenging2_hoeveelheid', 'bijmenging2_grondsoort',
-                'bijmenging3_plaatselijk', 'bijmenging3_hoeveelheid',
-                'bijmenging3_grondsoort']
+    df_default_columns = ['pkey_interpretatie', 'pkey_boring',
+                          'betrouwbaarheid_interpretatie', 'x', 'y',
+                          'diepte_laag_van', 'diepte_laag_tot',
+                          'hoofdnaam1_grondsoort', 'hoofdnaam2_grondsoort',
+                          'bijmenging1_plaatselijk', 'bijmenging1_hoeveelheid',
+                          'bijmenging1_grondsoort', 'bijmenging2_plaatselijk',
+                          'bijmenging2_hoeveelheid', 'bijmenging2_grondsoort',
+                          'bijmenging3_plaatselijk', 'bijmenging3_hoeveelheid',
+                          'bijmenging3_grondsoort']
 
     def test_search_nan(self, mp_wfs, mp_get_schema,
                         mp_remote_describefeaturetype, mp_remote_md,
@@ -101,8 +89,8 @@ class TestGecodeerdeLithologieSearch(AbstractTestSearch):
             Monkeypatch the call to get the remote XML data.
 
         """
-        df = self.get_search_object().search(
-            query=self.get_valid_query_single())
+        df = self.search_instance.search(
+            query=self.valid_query_single)
 
     def test_search_customreturnfields(self, mp_get_schema,
                                        mp_remote_describefeaturetype,
[…]

[…]
