### 2020-05-28 1dbb1f5
Add contributing installation instructions
