### 2020-05-28 e3b7c99
Refactor AbstractTestTypes to use class variables instead of getters.
