### 2020-05-28 7fb6539
Minor changes in wording.
