### 2020-05-28 b5562e0
Split development guidelines and contributing
