+++ .github/CONTRIBUTING.rst
@@ -11,7 +11,6 @@ rewarding for us - the project :ref:`authors` - to work on pydov.
 
 pydov is an open source project, maintained by people who care. We are not directly funded to do so.
 
-
 Code of conduct
 ---------------
 
@@ -126,230 +125,7 @@ Contribute code
 
 Care to fix bugs or implement new functionality for pydov? Awesome! Have a
 look at the `issue list`_ and leave a comment on the things you want
-to work on. See also the development guidelines below.
+to work on.
 
-.. _dev-guidelines:
-
-Development guidelines
------------------------
-
-Coding guidelines
-^^^^^^^^^^^^^^^^^^
-
-The following are some guidelines on how new code should be written. Of course,
-there are special cases and there will be exceptions to these rules. However,
-following these rules when submitting new code makes the review easier so new
-code can be integrated in less time.
-
-Uniformly formatted code makes it easier to share code ownership. The
-pydov project tries to closely follow the official Python guidelines
-detailed in `PEP8 <https://www.python.org/dev/peps/pep-0008/>`_ which detail
-how code should be formatted and indented. Please read it and follow it.
-
-In addition, we add the following guidelines:
-
-* Use underscores to separate words in non class names: ``n_samples`` rather than ``nsamples``.
-* Avoid multiple statements on one line. Prefer a line return after a control flow statement (\ ``if/for``\ ).
-* Please don’t use ``import *`` in any case. It is considered harmful by the official Python recommendations. It makes the code harder to read as the origin of symbols is no longer explicitly referenced, but most important, it prevents using a static analysis tool like pyflakes to automatically find bugs.
-* Use the `numpy docstring standard`_ in all your docstrings.
-* The attributes for specific classes are Pandas data.frames, please use lowercase names (eventually with `_`) as column names.
-
-
-Contribute to the repository
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-The preferred workflow for contributing is to fork the `main repository <https://github.com/DOV-Vlaanderen/pydov>`_ on GitHub, clone locally, and develop on a branch. For more information on this workflow, see the `github workflow <https://guides.github.com/introduction/flow/>`_.
-
-The workflow is provided for command line usage and using the `Github for Desktop <https://desktop.github.com/>`_ application. Feel free to use the environment you like the most.
-
-
-#.
-   Fork the `project repository <https://github.com/DOV-Vlaanderen/pydov>`_ by clicking on the 'Fork' button near the top right of the page. This creates a copy of the code under your personal GitHub user account.
-
-
-   .. image:: https://github-images.s3.amazonaws.com/help/bootcamp/Bootcamp-Fork.png
-      :target: https://github-images.s3.amazonaws.com/help/bootcamp/Bootcamp-Fork.png
-      :alt: forkrepo
-      :height: 200px
-
-
-#.
-   You’ve successfully forked the pydov repository, but so far, it only exists on GitHub. To be able to work on the project, you will need to clone it to your computer.
-
-    Clone your fork of the pydov repo from your GitHub account to your local disk:
-
-   .. code-block:: bash
-
-       $ git clone https://github.com/DOV-Vlaanderen/pydov.git
-       $ cd pydov
-
-   If you’re using the GitHub for Desktop application, navigate over to the bottom on the right hand side bar and click ``Clone in Desktop``. Once you've clicked this, it’ll ask you if you want to launch our desktop application to clone the repository, and where you want to save it. Pick a location on your computer that you feel comfortable with creating files and folders.
-
-
-   .. image:: https://guides.github.com/activities/forking/clone-in-desktop.png
[…]
