### 2020-05-28 a6cf515
Add link to contributing guidelines
