### 2020-05-28 a481989
Update package contributor installation instructions
