### 2020-05-28 DOV-Vlaanderen/pydov#256 @Roel
Fixed in #272
