### 2020-05-28 c083ec5
Fix #257 fix #259 and fix #275
