### 2020-05-29 5f66c2b
Add pandoc installation instruction
