### 2020-05-29 dd1fb0f
Add pandoc reference in "creating the docs" section.
