+++ docs/development.rst
@@ -170,7 +170,7 @@ documents, tutorials, etc. Documentation lives in the ``docs/`` directory.
 
 You can edit the documentation using any text editor and then generate the HTML
 output by typing ``make html`` from the ``doc/`` directory. For building the
-documentation, you will need `Sphinx`_. The ``_build``
+documentation, you will need `Sphinx`_ and `pandoc`_. The ``_build``
 directory is not included in the repository as we rely on CI tooling for the
 documentation building. The documentation is checked on Travis_ and build
 by `Read the docs`_.
@@ -222,3 +222,4 @@ The new release can be installed using ``pip``, ``pip install --upgrade pydov``.
 .. _packaging instructions: https://packaging.python.org/tutorials/packaging-projects/
 .. _numpy docstring standard: https://numpydoc.readthedocs.io/en/latest/format.html
 .. _Sphinx: http://www.sphinx-doc.org/en/master/
+.. _pandoc: https://pandoc.org/
