### 2020-05-29 DOV-Vlaanderen/pydov#255 @Roel
Closed in #274
