### 2020-05-29 4f25bb6
Readd full docstring in GrondwaterFilterSearch.
