+++ pydov/search/grondwaterfilter.py
@@ -83,7 +83,7 @@ class GrondwaterFilterSearch(AbstractSearch):
         """Search for objects of this type. Provide `location` and/or
         `query` and/or `max_features`.
         When `return_fields` is None, all fields are returned.
-        
+
         Excludes 'empty' filters (i.e. Putten without Filters) by extending
         the `query` with a not-null check on pkey_filter.
 
