### 2020-05-29 5812c12
Use global timeout instead of owslib default.
