### 2020-05-29 82bac38
Tryout setup conda environment
