+++ .github/workflows/ci.yml
@@ -15,11 +15,10 @@ jobs:
     runs-on: ${{ matrix.os }}
     
     strategy:
+      fail-fast: false
       matrix:
         os: [ubuntu-latest]
         python-version: ["3.5", "3.6", "3.7", "3.8"]
-        fail-fast: [false]
-          
     steps:
       - uses: "actions/checkout@v2"
       - uses: "actions/setup-python@v1"
@@ -44,11 +43,10 @@ jobs:
     runs-on: ${{ matrix.os }}
     
     strategy:
+      fail-fast: false
       matrix:
         os: [windows-latest]
         python-version: ["3.5", "3.6", "3.7", "3.8"]
-        fail-fast: [false]
-
     steps:
       - uses: "actions/checkout@v2"
       - uses: "actions/setup-python@v1"
@@ -66,3 +64,31 @@ jobs:
 
       - name: "Run tox targets for ${{ matrix.python-version }}"
         run: "python -m tox"
+
+  tests-conda:
+    name: "Python with conda (${{ matrix.python-version }}, ${{ matrix.os }})"
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
+        python-version: ["3.6", "3.7"]
+    steps:
+      - uses: "actions/checkout@v2"
+      - uses: goanpeca/setup-miniconda@v1
+        with:
+          channels: conda-forge
+          allow-softlinks: true
+          channel-priority: flexible
+          auto-update-conda: true
+          python-version: ${{ matrix.python-version }}
+      - name: Conda info
+        shell: bash
+        run: conda info
+      - name: Conda list
+        shell: pwsh
+        run: conda list
+
+
+
+
