### 2020-05-29 37dfe51
Add unit test step to conda
