### 2020-05-29 d3d5ada
Switch to fail fast false again
