+++ .github/workflows/ci.yml
@@ -15,7 +15,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     
     strategy:
-      fail-fast: true
+      fail-fast: false
       matrix:
         os: [ubuntu-latest]
         python-version: ["3.5", "3.6", "3.7", "3.8"]
@@ -43,7 +43,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     
     strategy:
-      fail-fast: true
+      fail-fast: false
       matrix:
         os: [windows-latest]
         python-version: ["3.5", "3.6", "3.7", "3.8"]
@@ -72,7 +72,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ["ubuntu-latest", "macos-latest", "windows-latest"]
-        python-version: ["3.6", "3.7"]
+        python-version: ["3.5", "3.6", "3.7", "3.8"]
     steps:
       - uses: "actions/checkout@v2"
       - uses: goanpeca/setup-miniconda@v1
