### 2020-05-29 7ebe4d8
Test with conda for linux and windows
