+++ .github/workflows/ci.yml
@@ -66,13 +66,13 @@ jobs:
         run: "python -m tox"
 
   tests-conda:
-    name: "Python with conda (${{ matrix.python-version }}, ${{ matrix.os }})"
+    name: "Python with conda (${{ matrix.python-version }} on ${{ matrix.os }})"
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
-        python-version: ["3.5", "3.6", "3.7", "3.8"]
+        os: ["ubuntu-latest", "windows-latest"]
+        python-version: ["3.5", "3.7", "3.8"]
     steps:
       - uses: "actions/checkout@v2"
       - uses: goanpeca/setup-miniconda@v1
@@ -86,6 +86,7 @@ jobs:
       - name: Install dependencies and package
         shell: pwsh
         run: |
+          conda update -n base -c defaults conda
           conda install --name pydov --file requirements.txt
           pip install -e .[devs]
           conda list
