### 2020-05-29 145b57a
Switch to python pytest command
