### 2020-05-29 9ce5f5f
Tryout fix Conda CI owslib
