+++ .github/workflows/ci.yml
@@ -90,7 +90,7 @@ jobs:
           conda install --name pydov --file requirements.txt
           conda list
       - name: adjust owslib version for py3.5
-        if: matrix.python-version == "3.5"
+        if: matrix.python-version == 3.5
         run: |
           conda install owslib=0.19.1
       - name: Install development dependencies with pip
