### 2020-05-29 7d06e03
Define conda env
