+++ .github/workflows/ci.yml
@@ -92,7 +92,7 @@ jobs:
       - name: adjust owslib version for py3.5
         if: matrix.python-version == 3.5
         run: |
-          conda install owslib=0.19.1
+          conda install --name pydov owslib=0.19.1
       - name: Install development dependencies with pip
         shell: pwsh
         run: |
