### 2020-06-03 DOV-Vlaanderen/pydov#254 @Roel
@pjhaest Did you make any progress on this yet? This was one of the main points in the pyOpenSci review so I think we need to tackle this before proceeding with the review process. Be sure to push your WIP so we can collaborate here :)
