+++ docs/index.rst
@@ -1,11 +1,8 @@
 Welcome to pydov's documentation!
 ==================================
 
-.. image:: https://travis-ci.org/DOV-Vlaanderen/pydov.svg?branch=master
-    :target: https://travis-ci.org/DOV-Vlaanderen/pydov
-
-.. image:: https://ci.appveyor.com/api/projects/status/4ljy2a0p661v3d9k/branch/master?svg=true
-    :target: https://ci.appveyor.com/project/Roel/pydov
+.. image:: https://github.com/DOV-Vlaanderen/pydov/workflows/CI/badge.svg
+    :target: https://github.com/DOV-Vlaanderen/pydov
 
 .. image:: https://readthedocs.org/projects/pydov/badge/?version=latest
     :target: http://pydov.readthedocs.io/en/latest/?badge=latest
