### 2020-06-15 DOV-Vlaanderen/pydov#254 @pjhaest
yes, sorry, last-minute project proposal came in between, along with some other things, as always.
I will start on it right away, if you haven't already? I was planning to add the tutorial as a seperate page to the docs, or would you prefer a seperate notebook?
