### 2020-06-22 DOV-Vlaanderen/pydov#254 @pjhaest
short update, I just started a new Notebook. Will upload tomorrow. Easier to comment/change than an rst. And probably even better to leave it at that and refer to the notebook in the docs.
