### 2020-06-29 f734149
gw vergunningen type refactored
