### 2020-07-04 1cff9ee
Update tutorials.rst
