+++ docs/tutorials.rst
@@ -7,8 +7,8 @@ Tutorials
 
 These tutorials illustrate use cases the pydov package supports, split
 up by different topics each setup as a `jupyter notebook`_
-you can run yourself. The notebook `Introductory_tutorial` bundles most 
-of the interesting features and methods of pydov. That would be a good
+you can run yourself. The introductory tutorial bundles most 
+of the interesting features and methods of pydov and would be a good
 start.
 
 .. _jupyter notebook: http://jupyter.org/
@@ -34,7 +34,7 @@ To run these interactively online without installation, use the following binder
 .. toctree::
    :maxdepth: 1
 
-   notebooks/Introductory_tutorial.ipynb
+   notebooks/introductory_tutorial.ipynb
    notebooks/search_boringen.ipynb
    notebooks/search_sonderingen.ipynb
    notebooks/search_grondwaterfilters.ipynb
