### 2020-07-16 5ddaac8
Add test for issue #285.
