### 2020-07-26 d21d3da
Update HISTORY for v2.0.0
