+++ HISTORY.rst
@@ -4,6 +4,50 @@
 History
 =======
 
+v2.0.0
+------
+
+* News
+
+  * This version drops support for Python 2.7 and Python 3.5.
+
+  * This version is supported on Python 3.6, 3.7 and 3.8.
+
+* Fixes and improvements
+
+  * Fix the korrelvolumemassa, volumemassa and watergehalte fields of Grondmonster type.
+
+  * Add the 'mv_mtaw' field to the GrondwaterFilter type.
+
+  * Extend the hooks system and distinguish between read and inject hooks. The 'xml_requested' hook has been removed in favor of 'xml_received'.
+
+  * Generate stable WFS GetFeature requests, allowing f.ex. hooks to reuse cached responses.
+
+* Development-only updates
+
+  * Remove some code duplication between pydov and OWSLib.
+
+  * Simplify test fixtures setup.
+
+  * Remove duplicate docstrings to simplify the codebase.
+
+* Documentation-only updates
+
+  * Add introductory tutorial.
+
+  * Add a tutorial on how to use a WFS geometry as location query.
+
+  * Update development installation instructions.
+
+  * Update folium examples to support the latest pyproj version.
+
+  * Add extra Binder links on top of each tutorial.
+
+  * Improve charts by including a title and axis labels.
+
+  * Improve README by adding dataframe output.
+
+
 v1.0.0
 ------
 
