### 2020-07-26 4ad107c
Remove Python3.5 from setup.py
