+++ setup.py
@@ -46,7 +46,6 @@ setup(
         'Natural Language :: English',
         'Natural Language :: Dutch',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
