### 2020-07-26 c2d0066
Bump version: 1.0.0 → 2.0.0
