+++ pydov/__init__.py
@@ -8,7 +8,7 @@ from pydov.util.hooks import (
 )
 
 __author__ = """DOV-Vlaanderen"""
-__version__ = '1.0.0'
+__version__ = '2.0.0'
 
 cache = pydov.util.caching.GzipTextFileCache()
 
