### 2020-07-27 #19 @Roel
Hi all! I hope everyone's doing well out there. We managed to tackle most of the raised points during our last code sprint and in the weeks that followed. There are still two issues unresolved for now, but we'll take these on in one of our next sprints for sure.

Let me know whether it suits you to pick back up here. Thanks for you patience and take care!
