+++ pydov/search/gwvergunningen.py
@@ -16,12 +16,7 @@ class GwvergunningenSearch(AbstractSearch):
     __wfs_namespace = None
     __md_metadata = None
     __xsd_schemas = None
-    # create dummy feature catalogue from custom definition in type
-    __fc_featurecatalogue = {'attributes': {}}
-    for wfs_field in Gwvergunningen.fields:
-        __fc_featurecatalogue['attributes'].update(
-            {wfs_field.source: {'definition': '', 'value': ''}}
-        )
+    __fc_featurecatalogue = None
 
     def __init__(self, objecttype=Gwvergunningen):
         """Initialisation.
@@ -49,24 +44,22 @@ class GwvergunningenSearch(AbstractSearch):
             if GwvergunningenSearch.__md_metadata is None:
                 GwvergunningenSearch.__md_metadata = self._get_remote_metadata()
 
-            # not present
-            # if GwvergunningenSearch.__fc_featurecatalogue is None:
-                #     csw_url = self._get_csw_base_url()
-                #     fc_uuid = owsutil.get_featurecatalogue_uuid(
-                #         GwvergunningenSearch.__md_metadata)
+            if GwvergunningenSearch.__fc_featurecatalogue is None:
+                csw_url = self._get_csw_base_url()
+                fc_uuid = owsutil.get_featurecatalogue_uuid(
+                    GwvergunningenSearch.__md_metadata)
 
-                # GwvergunningenSearch.__fc_featurecatalogue = \
-            #     owsutil.get_remote_featurecatalogue(csw_url, fc_uuid)
+                GwvergunningenSearch.__fc_featurecatalogue = \
+                owsutil.get_remote_featurecatalogue(csw_url, fc_uuid)
 
-            # if GwvergunningenSearch.__xsd_schemas is None:
-                #     GwvergunningenSearch.__xsd_schemas = \
-            #         self._get_remote_xsd_schemas()
+            if GwvergunningenSearch.__xsd_schemas is None:
+                GwvergunningenSearch.__xsd_schemas = \
+                    self._get_remote_xsd_schemas()
 
             fields = self._build_fields(
                 GwvergunningenSearch.__wfs_schema,
                 GwvergunningenSearch.__fc_featurecatalogue,
-                [], # GwvergunningenSearch.__xsd_schemas
-                source=('wfs',)
+                GwvergunningenSearch.__xsd_schemas
             )
 
             for field in fields.values():
@@ -78,9 +71,9 @@ class GwvergunningenSearch(AbstractSearch):
 
             self._fields = self._build_fields(
                 GwvergunningenSearch.__wfs_schema,
-                {'attributes': []}, # GwvergunningenSearch.__fc_featurecatalogue,
-                [], # GwvergunningenSearch.__xsd_schemas
-            )
+                GwvergunningenSearch.__fc_featurecatalogue,
+                GwvergunningenSearch.__xsd_schemas)
+
 
     def search(self, location=None, query=None, sort_by=None,
                return_fields=None, max_features=None):
