### 2020-08-14 ca0c2b6
addition of test files for GW vergunningen
