### 2020-08-14 6a919ed
change 'd' to 'diepte' for consistency
