### 2020-08-14 2f08b56
remove one duplicate attribute
