### 2020-08-14 bffb53b
write exceptions for the absence of XML in Gwvergunningen
