### 2020-10-14 #19 @lwasser
hi all. just checkin in to see whether all review comments have been addressed or who is working on what items at this point. many thanks.
