### 2020-10-19 d477c0c
default dataframe output in docs
