### 2020-10-19 6456a81
add type to reference in docs
