### 2020-10-19 380cf55
add tutorial to list of tutorials
