+++ docs/tutorials.rst
@@ -45,6 +45,7 @@ To run these interactively online without installation, use the following binder
    notebooks/search_geotechnische_codering.ipynb
    notebooks/search_quartaire_stratigrafie.ipynb
    notebooks/search_grondmonsters.ipynb
+   notebooks/search_gwvergunningen.ipynb
    notebooks/customizing_object_types.ipynb
    notebooks/remote_wfs_gml_query.ipynb
    notebooks/caching.ipynb
