### 2020-10-21 #19 @xmnlab
@Roel thank you so much for your patience! thanks for addressing the points I suggested.

@lwasser, `The author has responded to my review and made changes to my satisfaction. I recommend approving this package.`
