### 2020-10-21 #19 @NickleDave
Thank you for pinging me @lwasser -- sorry for losing track of this

Yes I see that https://github.com/DOV-Vlaanderen/pydov/issues/254 adds an introductory tutorial -- looks good @Roel, I think this can definitely help a newcomer quickly get up to speed on usage of `pydov`.

The other concern I raised about code duplication has been noted in https://github.com/DOV-Vlaanderen/pydov/issues/264 but I don't think this should be a blocking issue.

So, yes, the authors have responded and made changes to my satisfaction.
I recommend approving `pydov` (🎉)
