### 2020-10-26 DOV-Vlaanderen/pydov#264 @Roel
This might also relate to #281 as this will remove the major reason we implemented this shared data in the first place, which is to share the common (and large) WFS capabilities data between (instances of) search classes.
