### 2020-10-26 #19 @Roel
Awesome, this is great news! No worries about the timing, we're all busy :)

We will definetely pick up the remaining issues in one of our next sprints and keep them in mind when refactoring or working on new features.

Thank you again for your time @lwasser, @xmnlab and @NickleDave! Great to become a part of the growing list of pyOpenSci packages!
