### 2020-10-26 2914fef
added installation guide for Geopandas, Fiona and GDAL
