### 2020-10-26 dfd2fc7
updated the spatial_querying example to include a shape file conversion
