### 2020-10-27 #19 @NickleDave
Thank you for submitting `pydov` for review @Roel @pjhaest @stijnvanhoey  !!
I think it's a great tool, excited to tell more people about it.
Perfect for pyOpenSci. And I learned a lot from seeing you all develop it further.
