### 2020-10-28 #19 @lwasser
🎆  excellent! ok i think then we can consider pydov accepted as a part of pyopensci!!
thank you @Roel @pjhaest @stijnvanhoey and @NickleDave @xmnlab  for the efficient review process here!!

there are a few last steps if  someone is willing to take the time to do this

## Reviewers:

If you are not already on our website as a contributor, please submit a PR to [pyopensci.github.io](https://github.com/pyOpenSci/pyopensci.github.io) to add yourself as a [contributor](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/contributors.yml) using:

```
contributor_type:
    - reviewer
```

## Package Maintainer(s)

@stijnvanhoey  for you the tasks are:

- [x] Add the badge for pyOpenSci peer review, with the second link being the URL to this issue:
[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/19)
- [x] Submit a PR to `pyopensci.github.io` to add pydov to the pyOpenSci [package list](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/packages.yml) and to add yourself as a [contributor](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/contributors.yml) using:
```
contributor_type:
    - package-maintainer
```
and
```
packages-submitted: ["pydov"]
```
OPTIONAL!!

If you are interested (not required), you can write a blog post for the pyOpenSci website about PyDov (see blog post about [pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation)) to promote your package.  This is totally up to you!

Is the version that we are accepting here 2.0.0 ?
Thank you all!!
