### 2020-10-28 e607ad0
Add pyopensci badge to inde of documentation
