+++ docs/index.rst
@@ -10,6 +10,10 @@ Welcome to pydov's documentation!
 
 .. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.2788680.svg
    :target: https://doi.org/10.5281/zenodo.2788680
+   
+.. image:: https://camo.githubusercontent.com/63ff31cdb80a06361e53ac2b9ac0d184118ebd0b/68747470733a2f2f74696e7975726c2e636f6d2f7932326e62387570
+   :target: https://github.com/pyOpenSci/software-review/issues/19
+   :alt: pyopensci approval   
 
 A Python package to query and download data from Databank Ondergrond Vlaanderen (DOV).
 
