### 2020-10-28 b56ccee
added GeometryFilter which takes a geometry file type and tries to parse it to GML 3.1.1
