### 2020-10-28 78f5f4e
added GeometryFilter to pydov/util/location
