### 2020-11-07 e9328a4
Remove static variables and move common code into AbstractSearch.
