### 2020-11-08 aee11d7
Refactor out static variables in all search classes.
