+++ pydov/search/interpretaties.py
@@ -1,7 +1,5 @@
-import pandas as pd
-
+# -*- coding: utf-8 -*-
 from pydov.search.abstract import AbstractSearch
-from pydov.types.fields import _WfsInjectedField
 from pydov.types.interpretaties import (FormeleStratigrafie,
                                         GecodeerdeLithologie,
                                         GeotechnischeCodering,
@@ -10,17 +8,10 @@ from pydov.types.interpretaties import (FormeleStratigrafie,
                                         InformeleStratigrafie,
                                         LithologischeBeschrijvingen,
                                         QuartairStratigrafie)
-from pydov.util import owsutil
 
 
 class InformeleStratigrafieSearch(AbstractSearch):
     """Search class to retrieve information about 'informele stratigrafie'."""
-
-    __wfs_schema = None
-    __wfs_namespace = None
-    __md_metadata = None
-    __fc_featurecatalogue = None
-    __xsd_schemas = None
 
     def __init__(self, objecttype=InformeleStratigrafie):
         """Initialisation.
@@ -36,73 +27,10 @@ class InformeleStratigrafieSearch(AbstractSearch):
         super(InformeleStratigrafieSearch, self).__init__(
             'interpretaties:informele_stratigrafie', objecttype)
 
-    def _init_namespace(self):
-        if InformeleStratigrafieSearch.__wfs_namespace is None:
-            InformeleStratigrafieSearch.__wfs_namespace = self._get_namespace()
-
-    def _init_fields(self):
-        if self._fields is None:
-            if InformeleStratigrafieSearch.__wfs_schema is None:
-                InformeleStratigrafieSearch.__wfs_schema = self._get_schema()
-
-            if InformeleStratigrafieSearch.__md_metadata is None:
-                InformeleStratigrafieSearch.__md_metadata = \
-                    self._get_remote_metadata()
-
-            if InformeleStratigrafieSearch.__fc_featurecatalogue is None:
-                csw_url = self._get_csw_base_url()
-                fc_uuid = owsutil.get_featurecatalogue_uuid(
-                    InformeleStratigrafieSearch.__md_metadata)
-
-                InformeleStratigrafieSearch.__fc_featurecatalogue = \
-                    owsutil.get_remote_featurecatalogue(csw_url, fc_uuid)
-
-            if InformeleStratigrafieSearch.__xsd_schemas is None:
-                InformeleStratigrafieSearch.__xsd_schemas = \
-                    self._get_remote_xsd_schemas()
-
-            fields = self._build_fields(
-                InformeleStratigrafieSearch.__wfs_schema,
-                InformeleStratigrafieSearch.__fc_featurecatalogue,
-                InformeleStratigrafieSearch.__xsd_schemas)
-
-            for field in fields.values():
-                if field['name'] not in self._type.get_field_names(
-                        include_wfs_injected=True):
-                    self._type.fields.append(
-                        _WfsInjectedField(name=field['name'],
-                                          datatype=field['type']))
-
-            self._fields = self._build_fields(
-                InformeleStratigrafieSearch.__wfs_schema,
-                InformeleStratigrafieSearch.__fc_featurecatalogue,
-                InformeleStratigrafieSearch.__xsd_schemas)
-
-    def search(self, location=None, query=None, sort_by=None,
-               return_fields=None, max_features=None):
-        fts = self._search(location=location, query=query, sort_by=sort_by,
-                           return_fields=return_fields,
-                           extra_wfs_fields=['Type_proef', 'Proeffiche'],
-                           max_features=max_features)
-
-        interpretaties = self._type.from_wfs(
-            fts, self.__wfs_namespace)
-
[…]
