### 2020-11-08 b791366
Run Github CI on all branches.
