+++ pydov/__init__.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 import requests
+import urllib3
 
 import pydov.util.caching
-from pydov.util.hooks import (
-    SimpleStatusHook,
-    Hooks,
-)
+from pydov.util.hooks import Hooks, SimpleStatusHook
 
 __author__ = """DOV-Vlaanderen"""
 __version__ = '2.0.0'
@@ -20,4 +18,17 @@ hooks = Hooks(
 # session object allows connection pooling and TCP connection reuse.
 request_timeout = 300
 session = requests.Session()
+session.timeout = (60, 300)
+
 session.headers.update({'user-agent': 'pydov/{}'.format(pydov.__version__)})
+
+try:
+    retry = urllib3.util.Retry(total=10, connect=5, read=3, redirect=5,
+                               allowed_methods=set(['GET', 'POST']))
+except TypeError:
+    retry = urllib3.util.Retry(total=10, connect=5, read=3, redirect=5,
+                               method_whitelist=set(['GET', 'POST']))
+
+adapter = requests.adapters.HTTPAdapter(max_retries=retry)
+session.mount('http://', adapter)
+session.mount('https://', adapter)
