+++ pydov/search/abstract.py
@@ -111,24 +111,12 @@ class AbstractSearch(AbstractCommon):
         """
         if AbstractSearch.__wfs is None:
 
-            capabilities = HookRunner.execute_inject_meta_response(
-                build_dov_url('geoserver/wfs') + '?version=1.1.0'
-            )
+            capabilities = owsutil.get_url(build_dov_url(
+                'geoserver/wfs?request=GetCapabilities&version=1.1.0'))
 
-            if capabilities is None:
-                AbstractSearch.__wfs = WebFeatureService(
-                    url=build_dov_url('geoserver/wfs'), version="1.1.0",
-                    timeout=pydov.request_timeout)
-            else:
-                AbstractSearch.__wfs = WebFeatureService(
-                    url=build_dov_url('geoserver/wfs'), version="1.1.0",
-                    xml=capabilities)
-
-            HookRunner.execute_meta_received(
-                build_dov_url('geoserver/wfs') + '?version=1.1.0',
-                etree.tostring(AbstractSearch.__wfs._capabilities,
-                               encoding='utf8')
-            )
+            AbstractSearch.__wfs = WebFeatureService(
+                url=build_dov_url('geoserver/wfs'), version="1.1.0",
+                xml=capabilities)
 
     def _init_namespace(self):
         """Initialise the WFS namespace associated with the layer.
