### 2020-11-08 eff8844
Timout on Session object has no effect. Use custom HTTP adapter instead.
