### 2020-11-08 c5989b4
Use pydov session in update_test_data.py too.
