### 2020-11-18 275bee0
Add backoff_factor and use one session per thread.
