+++ pydov/util/dovutil.py
@@ -27,7 +27,7 @@ def build_dov_url(path):
     return base_url + path.lstrip('/')
 
 
-def get_remote_url(url):
+def get_remote_url(url, session=None):
     """Request the URL from the remote service and return its contents.
 
     Parameters
@@ -41,8 +41,10 @@ def get_remote_url(url):
         The raw XML data as bytes.
 
     """
+    if session is None:
+        session = pydov.session
 
-    request = pydov.session.get(url)
+    request = session.get(url)
     request.encoding = 'utf-8'
     return request.text.encode('utf8')
 
@@ -71,7 +73,7 @@ def get_xsd_schema(url):
     return response
 
 
-def get_dov_xml(url):
+def get_dov_xml(url, session=None):
     """Request the XML from the remote DOV webservices and return it.
 
     Parameters
@@ -88,7 +90,7 @@ def get_dov_xml(url):
     response = HookRunner.execute_inject_xml_response(url)
 
     if response is None:
-        response = get_remote_url(url)
+        response = get_remote_url(url, session)
 
     HookRunner.execute_xml_received(url, response)
 
