+++ .github/workflows/ci.yml
@@ -75,7 +75,7 @@ jobs:
         python-version: ["3.6", "3.7", "3.8"]
     steps:
       - uses: "actions/checkout@v2"
-      - uses: goanpeca/setup-miniconda@v1
+      - uses: conda-incubator/setup-miniconda@v2
         with:
           channels: conda-forge
           allow-softlinks: true
