### 2020-11-19 58091ab
Check service_ok once per test session only.
