+++ tox.ini
@@ -1,12 +1,11 @@
 [tox]
-envlist = {py35,py36,py37,py38}-{nolxml,lxml}, update-oefen, update-productie, flake8, docs
+envlist = {py36,py37,py38}-{nolxml,lxml}, update-oefen, update-productie, flake8, docs
 
 [gh-actions]
 python =
     3.8: py38, update-oefen, update-productie, docs, flake8
     3.7: py37
     3.6: py36
-    3.5: py35
 
 [testenv:flake8]
 basepython=python
@@ -26,7 +25,6 @@ commands=
 
 [testenv]
 basepython =
-    py35: python3.5
     py36: python3.6
     py37: python3.7
     py38: python3.8
