### 2020-11-19 b19f034
Use LocalSessionThreadPool in update_test_data too.
