+++ tests/data/update_test_data.py
@@ -1,6 +1,7 @@
 """Script to update the testdata based on DOV webservices."""
 import os
 import sys
+from multiprocessing import Lock
 
 from owslib.etree import etree
 
@@ -18,6 +19,9 @@ from pydov.types.interpretaties import (FormeleStratigrafie,
                                         QuartairStratigrafie)
 from pydov.types.sondering import Sondering
 from pydov.util.dovutil import build_dov_url, get_remote_url
+from pydov.util.net import LocalSessionThreadPool
+
+lock = Lock()
 
 
 def get_first_featuremember(wfs_response):
@@ -31,60 +35,77 @@ def get_first_featuremember(wfs_response):
             return etree.tostring(ft).decode('utf-8')
 
 
-def update_file(filepath, url, process_fn=None):
-    sys.stdout.write('Updating {} ...'.format(filepath))
+def update_file_real(filepath, url, process_fn=None, session=None):
+    output = 'Updated {} ... '.format(filepath)
     filepath = os.path.join(os.path.dirname(__file__), filepath)
     try:
-        data = get_remote_url(url)
+        data = get_remote_url(url, session)
         if isinstance(data, bytes):
             data = data.decode('utf-8')
     except Exception as e:
-        sys.stdout.write(' FAILED:\n   {}.\n'.format(e))
+        output += ' FAILED:\n   {}.\n'.format(e)
         return
     else:
         with open(filepath, 'wb') as f:
             if process_fn:
                 data = process_fn(data)
             f.write(data.encode('utf-8'))
-            sys.stdout.write(' OK.\n')
+            output += ' OK.\n'
+
+    with lock:
+        sys.stdout.write(output)
 
 
 if __name__ == '__main__':
-    # types/boring
+    pool = LocalSessionThreadPool()
 
+    def update_file(filepath, url, process_fn=None):
+        pool.execute(update_file_real, (filepath, url, process_fn))
+
+    # types/boring
     update_file('types/boring/boring.xml',
                 build_dov_url('data/boring/2004-103984.xml'))
 
-    update_file('types/boring/wfsgetfeature.xml',
-                build_dov_url('geoserver/ows?service=WFS'
-                '&version=1.1.0&request=GetFeature&typeName=dov-pub:Boringen'
-                '&maxFeatures=1&CQL_Filter=fiche=%27' + build_dov_url(
-                    'data/boring/2004-103984%27')))
+    update_file(
+        'types/boring/wfsgetfeature.xml',
+        build_dov_url(
+            'geoserver/ows?service=WFS'
+            '&version=1.1.0&request=GetFeature&typeName=dov-pub:Boringen'
+            '&maxFeatures=1&CQL_Filter=fiche=%27' +
+            build_dov_url('data/boring/2004-103984%27')))
 
-    update_file('types/boring/feature.xml',
-                build_dov_url('geoserver/ows?service=WFS'
-                '&version=1.1.0&request=GetFeature&typeName=dov-pub:Boringen'
-                '&maxFeatures=1&CQL_Filter=fiche=%27' + build_dov_url(
-                    'data/boring/2004-103984%27')),
-                get_first_featuremember)
+    update_file(
+        'types/boring/feature.xml',
+        build_dov_url(
+            'geoserver/ows?service=WFS'
+            '&version=1.1.0&request=GetFeature&typeName=dov-pub:Boringen'
+            '&maxFeatures=1&CQL_Filter=fiche=%27' +
+            build_dov_url('data/boring/2004-103984%27')),
+        get_first_featuremember)
 
[…]
