### 2020-11-21 c643129
Add reasonable timeout to GH actions.
