### 2020-11-21 5d6ed97
Split GH actions jobs between tests, updated data tests, flake8 and docs.
