+++ tox.ini
@@ -8,7 +8,7 @@ python =
     3.6: py36
 
 [testenv:flake8]
-basepython=python
+basepython=python3.8
 deps=flake8
 commands=flake8 pydov
 
