### 2020-11-21 a67c38f
Fix tox targets.
