+++ tox.ini
@@ -3,7 +3,7 @@ envlist = {py36,py37,py38}-{nolxml,lxml}, update-oefen, update-productie, flake8
 
 [gh-actions]
 python =
-    3.8: py38, update-oefen, update-productie, docs, flake8
+    3.8: py38
     3.7: py37
     3.6: py36
 
