+++ pydov/util/net.py
@@ -11,7 +11,12 @@ request_timeout = 300
 
 
 class TimeoutHTTPAdapter(HTTPAdapter):
+    """HTTPAdapter which adds a default timeout to requests. Allows timeout
+    do be overridden on a per-request basis.
+    """
+
     def __init__(self, *args, **kwargs):
+        """Initialisation."""
         self.timeout = request_timeout
         if "timeout" in kwargs:
             self.timeout = kwargs["timeout"]
@@ -19,6 +24,18 @@ class TimeoutHTTPAdapter(HTTPAdapter):
         super().__init__(*args, **kwargs)
 
     def send(self, request, **kwargs):
+        """Sends PreparedRequest object. Returns Response object.
+
+        Parameters
+        ----------
+        request : requests.PreparedRequest
+            The PreparedRequest being sent.
+
+        Returns
+        -------
+        requests.Response
+            The Respone of the request.
+        """
         timeout = kwargs.get("timeout")
         if timeout is None:
             kwargs["timeout"] = self.timeout
@@ -26,8 +43,21 @@ class TimeoutHTTPAdapter(HTTPAdapter):
 
 
 class SessionFactory:
+    """Class for generating pydov configured requests Sessions. They are used
+    to send HTTP requests using our user-agent and with added retry-logic.
+
+    One global session is generated for all requests, and additionally one
+    session is generated per thread executing XML requests in parallel.
+    """
     @staticmethod
     def get_session():
+        """Request a new session.
+
+        Returns
+        -------
+        requests.Session
+            pydov configured requests Session.
+        """
         session = requests.Session()
 
         session.headers.update(
@@ -50,7 +80,20 @@ class SessionFactory:
 
 
 class LocalSessionThreadPool:
+    """Thread pool of LocalSessionThreads used to perform XML I/O operations
+    in parallel.
+    """
+
     def __init__(self, workers=4):
+        """Initialisation.
+
+        Set up the pool and start all workers.
+
+        Parameters
+        ----------
+        workers : int, optional
+            Number of worker threads to use, defaults to 4.
+        """
         self.workers = []
         self.input_queue = Queue(maxsize=100)
         self.result_queue = Queue()
@@ -59,22 +102,45 @@ class LocalSessionThreadPool:
             self.workers.append(
                 LocalSessionThread(self.input_queue, self.result_queue))
 
-        self.start()
+        self._start()
 
-    def start(self):
+    def _start(self):
+        """Start all worker threads. """
         for w in self.workers:
             w.start()
 
     def stop(self):
+        """Stop all worker threads. """
         for w in self.workers:
             w.stop()
 
     def execute(self, fn, args):
+        """Execute the given function with its arguments in a worker thread.
+
+        This will add the job to the queue and will not wait for the result.
+        Use join() to retrieve the result.
+
+        Parameters
+        ----------
+        fn : function
+            Function to execute.
+        args : tuple
+            Arguments that will be passed to the function.
+        """
         r = WorkerResult()
         self.input_queue.put((fn, args, r))
         self.result_queue.put(r)
 
     def join(self):
+        """Wait for all the jobs to be executed and return the results of all
+        jobs in a list.
+
+        Returns
+        -------
+        list
+            List of the result of all executed function in the order they were
+            submitted.
+        """
         self.input_queue.join()
         self.stop()
 
@@ -85,29 +151,68 @@ class LocalSessionThreadPool:
 
 
 class WorkerResult:
+    """Class for storing the result of a job execution in the result queue.
+
+    This allows putting a result instance in the queue on job submission and
+    fill in the result later when the job completes. This ensures the result
+    output is in the same order as the jobs were submitted.
+    """
+
     def __init__(self):
[…]
