### 2020-11-21 e538afb
Fix thread initialisation.
