+++ tox.ini
@@ -1,8 +1,9 @@
 [tox]
-envlist = {py36,py37,py38}-{nolxml,lxml}, update-oefen, update-productie, flake8, docs
+envlist = {py36,py37,py38,py39}-{nolxml,lxml}, update-oefen, update-productie, flake8, docs
 
 [gh-actions]
 python =
+    3.9: py39
     3.8: py38
     3.7: py37
     3.6: py36
@@ -28,6 +29,7 @@ basepython =
     py36: python3.6
     py37: python3.7
     py38: python3.8
+    py39: python3.9
 setenv =
     PYTHONPATH = {toxinidir}
 deps =
@@ -61,8 +63,3 @@ commands =
     pip install -U pip
     python {toxinidir}/tests/data/update_test_data.py
     py.test --basetemp={envtmpdir} --cov=pydov
-
-; If you want to make tox run the tests with the same versions, create a
-; requirements.txt with the pinned versions and uncomment the following lines:
-; deps =
-;     -r{toxinidir}/requirements.txt
