### 2020-11-22 5f0f41b
Add XDOV to ServiceCheck.
