### 2020-11-22 1c65206
Only update test data if services are online.
