### 2020-12-22 c2f7e5f
Allow retrying HEAD requests.
