### 2020-12-22 82b436c
Patch OWSLib's function for getting the DescribeFeatureType.
