### 2020-12-22 7c14ea8
Revert previous DescribeFeatureType patch in favor of patching openURL completely.
