### 2021-01-08 f694f5a
Only patch openURL in test suite, not at runtime.
To avoid conflicts when using pydov together with OWSLib (f.ex. to call other services).
