### 2021-01-08 f913816
Try to fix openURL monkeypatch.
