### 2021-01-11 #19 @lwasser
Hi  there @stijnvanhoey ! i'm doing some housekeeping today and i noticed pydov is [not on our list of packages here](https://www.pyopensci.org/python-packages/). Nor are you listed as a contributor!  did a PR get missed? i don't see one but that  does not mean i didn't miss something. we can close this once  this step is complete! i hope you are well!
