### 2021-02-18 486a10c
Attributes bodemlocatie
