+++ tests/data/update_test_data.py
@@ -5,6 +5,7 @@ from multiprocessing import Lock
 
 from owslib.etree import etree
 
+from pydov.types.bodemlocatie import Bodemlocatie
 from pydov.types.boring import Boring
 from pydov.types.grondmonster import Grondmonster
 from pydov.types.grondwaterfilter import GrondwaterFilter
@@ -821,5 +822,53 @@ if __name__ == '__main__':
         update_file(
             'types/grondmonster/xsd_%s.xml' %
             xsd_schema.split('/')[-1], xsd_schema)
+
+    # types/bodemlocatie
+    update_file('types/bodemlocatie/bodemlocatie.xml',
+                build_dov_url('data/bodemlocatie/2011-000002.xml'))
+
+    update_file(
+        'types/bodemlocatie/wfsgetfeature.xml',
+        build_dov_url(
+            'geoserver/ows?service=WFS'
+            '&version=1.1.0&request=GetFeature&typeName=bodem:bodemlocaties'
+            '&maxFeatures=1&CQL_Filter=Bodemlocatiefiche=%27' +
+            build_dov_url('data/bodemlocatie/2011-000002%27')))
+
+    update_file(
+        'types/bodemlocatie/feature.xml',
+        build_dov_url(
+            'geoserver/ows?service=WFS'
+            '&version=1.1.0&request=GetFeature&typeName=bodem:bodemlocaties'
+            '&maxFeatures=1&CQL_Filter=Bodemlocatiefiche=%27' +
+            build_dov_url('data/bodemlocatie/2011-000002%27')),
+        get_first_featuremember)
+
+    update_file(
+        'types/bodemlocatie/fc_featurecatalogue.xml',
+        build_dov_url(
+            'geonetwork/srv/dut/csw'
+            '?Service=CSW&Request=GetRecordById&Version=2.0.2'
+            '&outputSchema=http://www.isotc211.org/2005/gfc'
+            '&elementSetName=full&id=89d4f9a1-0474-4ade-b30f-442c31d17dc6'))
+
+    update_file(
+        'types/bodemlocatie/md_metadata.xml',
+        build_dov_url(
+            'geonetwork/srv/dut/csw'
+            '?Service=CSW&Request=GetRecordById&Version=2.0.2'
+            '&outputSchema=http://www.isotc211.org/2005/gmd'
+            '&elementSetName=full&id=3f507fd9-24c0-40ab-9328-29f0dff571fe'))
+
+    update_file(
+        'types/bodemlocatie/wfsdescribefeaturetype.xml',
+        build_dov_url(
+            'geoserver/bodem/bodemlocaties'
+            '/ows?service=wfs&version=1.1.0&request=DescribeFeatureType'))
+
+    for xsd_schema in Bodemlocatie.get_xsd_schemas():
+        update_file(
+            'types/bodemlocatie/xsd_{}.xml'.format(xsd_schema.split('/')[-1]),
+            xsd_schema)
 
     pool.join()

[…]
