### 2021-02-18 57c9757
Update installation instructions
