### 2021-02-18 d224ae9
Update bodemobservatie.py
