+++ pydov/types/boring.py
@@ -74,7 +74,7 @@ class Boring(AbstractDovType):
             `https://www.dov.vlaanderen.be/data/boring/<id>`.
 
         """
-        super(Boring, self).__init__('boring', pkey)
+        super().__init__('boring', pkey)
 
     @classmethod
     def from_wfs_element(cls, feature, namespace):
