+++ tests/test_error_path.py
@@ -1,17 +1,23 @@
+import datetime
+import gzip
 import os
 import platform
 import signal
 import sys
 import tempfile
 import time
+from importlib import reload
 from subprocess import Popen
 
+import numpy as np
 import pytest
 from owslib.fes import PropertyIsEqualTo
 
 import pydov
 from pydov.search.boring import BoringSearch
+from pydov.search.grondwaterfilter import GrondwaterFilterSearch
 from pydov.util.caching import GzipTextFileCache
+from pydov.util.errors import XmlFetchWarning, XmlStaleWarning, XsdFetchWarning
 from tests.abstract import ServiceCheck
 
 
@@ -23,13 +29,9 @@ def dov_proxy_no_xdov():
     time.sleep(2)
 
     os.environ['PYDOV_BASE_URL'] = 'http://localhost:1337/'
-    gziptext_cache = GzipTextFileCache(
-        cachedir=os.path.join(tempfile.gettempdir(), 'pydov_tests_error'))
-    pydov.cache = gziptext_cache
 
-    yield gziptext_cache
+    yield
 
-    gziptext_cache.remove()
     del os.environ['PYDOV_BASE_URL']
 
     if platform.system() == 'Windows':
@@ -39,6 +41,46 @@ def dov_proxy_no_xdov():
     time.sleep(2)
 
 
+@pytest.fixture(scope="module", autouse=True)
+def reload_modules(dov_proxy_no_xdov):
+    reload(pydov.types.boring)
+    reload(pydov.types.grondwaterfilter)
+
+    yield
+
+    reload(pydov.types.boring)
+    reload(pydov.types.grondwaterfilter)
+
+
+@pytest.fixture(scope="function", autouse=True)
+def reset_cache(dov_proxy_no_xdov):
+    """Reset the cache to a temporary folder to remove influence from other
+    tests.
+
+    The cache needs to be reset after setting the PYDOV_BASE_URL variable
+    because at initialisation this URL is used to construct a regex for
+    determining the datatype of an XML request.
+
+    Parameters
+    ----------
+    dov_proxy_no_xdov : pytest.fixture
+        Fixture starting up the DOV proxy stub and setting PYDOV_BASE_URL
+        accordingly.
+    """
+    gziptext_cache = GzipTextFileCache(
+        cachedir=os.path.join(tempfile.gettempdir(), 'pydov_tests_error'),
+        max_age=datetime.timedelta(seconds=0.1))
+    gziptext_cache.remove()
+
+    orig_cache = pydov.cache
+    pydov.cache = gziptext_cache
+
+    yield
+
+    gziptext_cache.remove()
+    pydov.cache = orig_cache
+
+
 class TestErrorPaths(object):
     """Class grouping tests related failing DOV services."""
 
@@ -46,12 +88,112 @@ class TestErrorPaths(object):
     @pytest.mark.skipif(not ServiceCheck.service_ok(),
                         reason="DOV service is unreachable")
     def test_do_not_cache_error(self):
-        bs = BoringSearch()
+        bs = BoringSearch(objecttype=pydov.types.boring.Boring)
+
         bs.search(query=PropertyIsEqualTo(
+            'pkey_boring',
+            'https://www.dov.vlaanderen.be/data/boring/2004-103984'))
+
+        assert not os.path.exists(os.path.join(
+            pydov.cache.cachedir, 'boring', '2004-103984.xml.gz'
+        ))
+        assert False
+
+    @pytest.mark.online
+    @pytest.mark.skipif(not ServiceCheck.service_ok(),
+                        reason="DOV service is unreachable")
+    def test_do_not_overwrite_stale_cache(self):
+        bs = BoringSearch(objecttype=pydov.types.boring.Boring)
+
[…]
