+++ pydov/types/grondmonster.py
@@ -143,7 +143,7 @@ class Grondmonster(AbstractDovType):
             `https://www.dov.vlaanderen.be/data/grondmonster/<id>`.
 
         """
-        super(Grondmonster, self).__init__('grondmonster', pkey)
+        super().__init__('grondmonster', pkey)
 
     @classmethod
     def from_wfs_element(cls, feature, namespace):
