+++ pydov/types/sondering.py
@@ -93,7 +93,7 @@ class Sondering(AbstractDovType):
             the form `https://www.dov.vlaanderen.be/data/sondering/<id>`.
 
         """
-        super(Sondering, self).__init__('sondering', pkey)
+        super().__init__('sondering', pkey)
 
     @classmethod
     def from_wfs_element(cls, feature, namespace):
