+++ tests/test_search_bodemlocatie.py
@@ -1,4 +1,4 @@
-"""Module grouping tests for the boring search module."""
+"""Module grouping tests for the bodemlocatie search module."""
 import datetime
 
 from owslib.fes import PropertyIsEqualTo
@@ -7,41 +7,38 @@ from pydov.search.bodemlocatie import BodemlocatieSearch
 from pydov.types.bodemlocatie import Bodemlocatie
 from tests.abstract import AbstractTestSearch
 
-location_md_metadata = 'tests/data/types/boring/md_metadata.xml'
+location_md_metadata = 'tests/data/types/bodemlocatie/md_metadata.xml'
 location_fc_featurecatalogue = \
-    'tests/data/types/boring/fc_featurecatalogue.xml'
+    'tests/data/types/bodemlocatie/fc_featurecatalogue.xml'
 location_wfs_describefeaturetype = \
-    'tests/data/types/boring/wfsdescribefeaturetype.xml'
-location_wfs_getfeature = 'tests/data/types/boring/wfsgetfeature.xml'
-location_wfs_feature = 'tests/data/types/boring/feature.xml'
-location_dov_xml = 'tests/data/types/boring/boring.xml'
-location_xsd_base = 'tests/data/types/boring/xsd_*.xml'
+    'tests/data/types/bodemlocatie/wfsdescribefeaturetype.xml'
+location_wfs_getfeature = 'tests/data/types/bodemlocatie/wfsgetfeature.xml'
+location_wfs_feature = 'tests/data/types/bodemlocatie/feature.xml'
+location_dov_xml = 'tests/data/types/bodemlocatie/bodemlocatie.xml'
+location_xsd_base = 'tests/data/types/bodemlocatie/xsd_*.xml'
 
 
-class TestBoringSearch(AbstractTestSearch):
+class TestBodemlocatieSearch(AbstractTestSearch):
 
-    search_instance = BoringSearch()
-    datatype_class = Boring
+    search_instance = BodemlocatieSearch()
+    datatype_class = Bodemlocatie
 
-    valid_query_single = PropertyIsEqualTo(propertyname='boornummer',
-                                           literal='GEO-04/169-BNo-B1')
+    valid_query_single = PropertyIsEqualTo(propertyname='naam',
+                                           literal='STARC_4')
 
     inexistent_field = 'onbestaand'
-    wfs_field = 'boornummer'
-    xml_field = 'boormethode'
+    wfs_field = 'naam'
+    xml_field = 'waarnemingsdatum'
 
-    valid_returnfields = ('pkey_boring', 'boornummer', 'diepte_boring_tot',
-                          'datum_aanvang')
-    valid_returnfields_subtype = ('pkey_boring', 'boornummer',
-                                  'diepte_methode_van', 'diepte_methode_tot')
-    valid_returnfields_extra = ('pkey_boring', 'doel')
+    valid_returnfields = ('pkey_bodemlocatie', 'naam', 'waarnemingsdatum')
+    valid_returnfields_subtype = ('pkey_bodemlocatie', 'naam', 'waarnemingsdatum')
+    valid_returnfields_extra = ('pkey_bodemlocatie', 'naam')
 
-    df_default_columns = ['pkey_boring', 'boornummer', 'x', 'y', 'mv_mtaw',
-                          'start_boring_mtaw', 'gemeente',
-                          'diepte_boring_van', 'diepte_boring_tot',
-                          'datum_aanvang', 'uitvoerder', 'boorgatmeting',
-                          'diepte_methode_van', 'diepte_methode_tot',
-                          'boormethode']
+
+    df_default_columns = ['pkey_bodemlocatie', 'naam', 'type', 'datum', 'doel', 'x', 'y', 'z',
[…]
