+++ tests/test_types.py
@@ -2,6 +2,7 @@
 
 import pytest
 
+from pydov.types.bodemsite import Bodemsite
 from pydov.types.bodemlocatie import Bodemlocatie
 from pydov.types.boring import Boring
 from pydov.types.fields import XmlField
@@ -20,7 +21,8 @@ from pydov.types.interpretaties import (
 )
 from pydov.types.sondering import Sondering
 
-type_objects = [Bodemlocatie,
+type_objects = [Bodemsite,
+                Bodemlocatie,
                 Boring,
                 Sondering,
                 GrondwaterFilter,
