+++ tests/test_search.py
@@ -2,6 +2,7 @@
 
 import pytest
 
+from pydov.search.bodemsite import BodemsiteSearch
 from pydov.search.bodemlocatie import BodemlocatieSearch
 from pydov.search.boring import BoringSearch
 from pydov.search.grondmonster import GrondmonsterSearch
@@ -17,7 +18,8 @@ from pydov.util.errors import InvalidSearchParameterError
 from pydov.util.location import Point, WithinDistance
 from tests.abstract import ServiceCheck
 
-search_objects = [BodemlocatieSearch(),
+search_objects = [BodemsiteSearch(),
+                  BodemlocatieSearch(),
                   BoringSearch(),
                   SonderingSearch(),
                   GrondwaterFilterSearch(),
