### 2021-02-19 046be1e
update_test_data for wfs response to utf-8
