### 2021-02-19 bafff79
Define exceptions on invalid geopandas input
