### 2021-02-19 19f42b8
Add unit tests on Geometry and Geopandas filters
