+++ tests/test_util_location_gmlfilter.py
@@ -6,6 +6,7 @@ from owslib.fes import (
 from pydov.util.location import (
     GmlFilter,
     GeometryFilter,
+    GeopandasFilter,
     WithinDistance,
     Within,
     Disjoint,
@@ -14,6 +15,8 @@ from pydov.util.location import (
 from owslib.etree import etree
 from pydov.util.owsutil import set_geometry_column
 from tests.abstract import clean_xml
+
+import pytest
 
 
 class TestPoint(object):
@@ -742,3 +745,105 @@ class TestCombination(object):
             '186910.000000</gml:lowerCorner><gml:upperCorner>112220.000000 '
             '202870.000000</gml:upperCorner></gml:Envelope></ogc'
             ':Within></ogc:Or>')
+
+
+class TestGeometry(object):
+
+    def test_shapefile(self):
+        """Test the conversion of a vector file to GML using fiona,
+        Within operator.
+
+        Test whether the generated XML is correct and stable.
+        """
+        shapefile = 'tests/data/util/location/polygon_multiple_31370.shp'
+
+        f = GeometryFilter(shapefile, Within)
+        f.set_geometry_column('geom')
+        xml = f.toXML()
+
+        assert clean_xml(etree.tostring(xml).decode('utf8')) == clean_xml(
+            '<ogc:Or xmlns:gml311="http://www.opengis.net/gml" '
+            'xmlns:ogc="http://www.opengis.net/ogc"><ogc:Within><ogc:'
+            'PropertyName>geom</ogc:PropertyName><gml311:Polygon srsName='
+            '"urn:ogc:def:crs:EPSG::31370"><gml311:exterior><gml311:LinearRing>'
+            '<gml311:posList>108636.150020818 194960.844295764 109195.573506438 '
+            '195118.42837622 108911.922161617 194291.111953824 108636.150020818 '
+            '194960.844295764</gml311:posList></gml311:LinearRing></gml311:exterior>'
+            '</gml311:Polygon></ogc:Within><ogc:Within><ogc:PropertyName>'
+            'geom</ogc:PropertyName><gml311:Polygon srsName="urn:ogc:def:crs:'
+            'EPSG::31370"><gml311:exterior><gml311:LinearRing><gml311:posList>'
+            '107485.786233486 196741.544404921 108297.344247837 196843.974057217 '
+            '107840.350414513 196339.704999757 107485.786233486 196741.544404921'
+            '</gml311:posList></gml311:LinearRing></gml311:exterior></gml311:Polygon>'
+            '</ogc:Within></ogc:Or>')
+
+    def test_geopandas_dataframe(self):
+        """Test the conversion of a GeoPandas GeoDataFrame
+
+        Test whether the generated XML is correct and stable.
+        """
+        shapefile = 'tests/data/util/location/polygon_multiple_31370.shp'
+        import geopandas as gpd
+        gdf = gpd.read_file(shapefile)
+
+        f = GeopandasFilter(gdf, Within)
+        f.set_geometry_column('geom')
+        xml = f.toXML()
+
+        assert clean_xml(etree.tostring(xml).decode('utf8')) == clean_xml(
+            '<ogc:Or xmlns:gml311="http://www.opengis.net/gml" '
+            'xmlns:ogc="http://www.opengis.net/ogc"><ogc:Within><ogc:'
+            'PropertyName>geom</ogc:PropertyName><gml311:Polygon srsName='
+            '"urn:ogc:def:crs:EPSG::31370"><gml311:exterior><gml311:LinearRing>'
+            '<gml311:posList>108636.150020818 194960.844295764 109195.573506438 '
+            '195118.42837622 108911.922161617 194291.111953824 108636.150020818 '
[…]
