+++ tox.ini
@@ -35,6 +35,8 @@ setenv =
 deps =
     -r{toxinidir}/requirements_dev.txt
     lxml: lxml
+    fiona
+    geopandas
 commands =
     py.test --basetemp={envtmpdir} --cov=pydov
 
