+++ tox.ini
@@ -35,8 +35,6 @@ setenv =
 deps =
     -r{toxinidir}/requirements_dev.txt
     lxml: lxml
-    fiona
-    geopandas
 commands =
     py.test --basetemp={envtmpdir} --cov=pydov
 
