+++ pydov/types/bodemobservatie.py
@@ -12,11 +12,15 @@ class Bodemobservatie(AbstractDovType):
     subtypes = []
 
     fields = [
-        WfsField(name='pkey_bodemlocatie', source_field='Bodemobservatiefiche', datatype='string'),
-        WfsField(name='parameter', source_field='Parameter', datatype='string'),
+        WfsField(name='pkey_bodemlocatie',
+                 source_field='Bodemobservatiefiche',
+                 datatype='string'),
+        WfsField(name='parameter', source_field='Parameter',
+                 datatype='string'),
         XmlField(name='parametergroep',
                  source_xpath='/bodemobservatie/parametergroep',
-                 definition='Indeling van de parameter naar groep, bvb anionen, kationen, .... Is indicatief.',
+                 definition='Indeling van de parameter naar groep, '
+                 + 'bvb anionen, kationen, .... Is indicatief.',
                  datatype='string'),
         WfsField(name='waarde', source_field='Waarde', datatype='string'),
         WfsField(name='eenheid', source_field='Eenheid', datatype='string'),
@@ -28,14 +32,18 @@ class Bodemobservatie(AbstractDovType):
                  source_xpath='/bodemobservatie/bovengrens',
                  definition='Het mogelijke maximum van de meting.',
                  datatype='float'),
-        WfsField(name='methode', source_field='Observatiemethode', datatype='string'),
+        WfsField(name='methode', source_field='Observatiemethode',
+                 datatype='string'),
         XmlField(name='betrouwbaarheid',
                  source_xpath='/bodemobservatie/betrouwbaarheid',
                  definition='Betrouwbaarheid van de meting.',
                  datatype='string'),
-        WfsField(name='veld_labo', source_field='Labo_of_veld', datatype='string'),
-        WfsField(name='diepte_van', source_field='Diepte_van', datatype='float'),
-        WfsField(name='diepte_tot', source_field='Diepte_tot', datatype='float'),
+        WfsField(name='veld_labo', source_field='Labo_of_veld',
+                 datatype='string'),
+        WfsField(name='diepte_van', source_field='Diepte_van',
+                 datatype='float'),
+        WfsField(name='diepte_tot', source_field='Diepte_tot',
+                 datatype='float'),
         XmlField(name='observatiedatum',
                  source_xpath='/bodemobservatie/observatiedatum',
                  definition='Observatiedatum van de bodemobservatie',
