### 2021-02-19 bcc639b
Checkstyle fixes
