### 2021-02-19 2156aab
Edit bodemobservatie configuration
