### 2021-02-19 3188617
Revert changes in search/abstract.py
