+++ docs/tutorials.rst
@@ -39,6 +39,7 @@ To run these interactively online without installation, use the following binder
    notebooks/search_sonderingen.ipynb
    notebooks/search_grondwaterfilters.ipynb
    notebooks/search_grondwatermonsters.ipynb
+   notebooks/search_grondwatervergunningen.ipynb
    notebooks/search_formele_stratigrafie.ipynb
    notebooks/search_informele_stratigrafie.ipynb
    notebooks/search_hydrogeologische_stratigrafie.ipynb
@@ -48,7 +49,6 @@ To run these interactively online without installation, use the following binder
    notebooks/search_geotechnische_codering.ipynb
    notebooks/search_quartaire_stratigrafie.ipynb
    notebooks/search_grondmonsters.ipynb
-   notebooks/search_gwvergunningen.ipynb
    notebooks/customizing_object_types.ipynb
    notebooks/remote_wfs_gml_query.ipynb
    notebooks/caching.ipynb
