### 2021-02-19 4fa8a1d
Add test for search without pkey in return fields.
