### 2021-02-19 f7bfa1b
Move test to test_search instead.
