### 2021-02-22 7f1d718
Add installation instructions of @rebot
