### 2021-02-22 b2cfc30
Fix merge issue: capabilities via pydov session.
