### 2021-02-23 c78645c
Update documentation for bodemmonster datatype
