+++ pydov/types/abstract.py
@@ -354,7 +354,7 @@ class AbstractDovType(AbstractTypeCommon):
                 XmlParseWarning)
             return False
 
-    @ classmethod
+    @classmethod
     def from_wfs_element(cls, feature, namespace):
         """Build an instance of this type from a WFS feature element.
 
@@ -380,7 +380,7 @@ class AbstractDovType(AbstractTypeCommon):
         """
         raise NotImplementedError('This should be implemented in a subclass.')
 
-    @ classmethod
+    @classmethod
     def from_wfs(cls, response, namespace):
         """Build instances of this type from a WFS response.
 
@@ -425,7 +425,7 @@ class AbstractDovType(AbstractTypeCommon):
             for el in response:
                 yield (cls.from_wfs_element(el, namespace))
 
-    @ classmethod
+    @classmethod
     def get_field_names(cls, return_fields=None, include_subtypes=True,
                         include_wfs_injected=False):
         """Return the names of the fields available for this type.
@@ -484,7 +484,7 @@ class AbstractDovType(AbstractTypeCommon):
                         "Unknown return field: '{}'".format(rf))
         return fields
 
-    @ classmethod
+    @classmethod
     def get_fields(cls, source=('wfs', 'xml'), include_subtypes=True):
         """Return the metadata of the fields available for this type.
 
@@ -543,7 +543,7 @@ class AbstractDovType(AbstractTypeCommon):
 
         return fields
 
-    @ classmethod
+    @classmethod
     def get_xsd_schemas(cls):
         """Get a set of distinct XSD schema URLs for this type and its
         subtypes.
@@ -564,7 +564,7 @@ class AbstractDovType(AbstractTypeCommon):
 
         return xsd_schemas
 
-    @ classmethod
+    @classmethod
     def to_df_array(cls, iterable, return_fields=None):
         """Returns a dataframe array with one or more arrays (rows) for each
         instance in the given iterable.
