### 2021-02-28 40104df
Yield WorkerResult instead of returning a list of results.
