### 2021-03-02 83c0c5f
Some changes in bodem-documentation after review
