### 2021-03-03 46dcd2b
Fixed Windows CI
Fixed Windows CI + Adapted the installation guidelines for GDAL, Fiona and GeoPandas
