### 2022-10-19 #19 @lwasser
actually i think there are three maintainers possible here?
@Roel , @pjhaest I just pinged @stijnvanhoey above but would appreciate your filling our our[🔗 SURVEY](https://forms.gle/jgPA8NedJwgnGGBMA) as well. and you are also welcome to join us on slack. Many thanks again for your time.
