### 2022-10-28 #19 @lwasser
Hey there - @Roel, @stijnvanhoey, @pjhaest just checking in again here to make sure you recieved my message above! It's important that packages in our pyOpenSci ecosystem are maintained and that we have some level of communication with our maintainers. I think pydov is still being maintained.

But i'd like to know who the current maintainers are so  please do get back to me here on this issue. And we'd GREATLY appreciate your time in filling out [this survey ](https://forms.gle/jgPA8NedJwgnGGBMA) as well.

Many thanks for your time in doing these things! 🙌
