### 2022-10-29 #19 @Roel
Hi @lwasser

I'm indeed doing most of the maintenance of pydov currently. We had a bugfix release just a few days ago and a new major release is coming up soon!

I hope to find the time next week to fill in the survey.
